% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapwhale-styler-funksjonar.R
\name{add_line_break_after_pipe_rapwhale}
\alias{add_line_break_after_pipe_rapwhale}
\title{Add line break after pipe}
\usage{
add_line_break_after_pipe_rapwhale(pd)
}
\arguments{
\item{pd}{A parse table.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Legg til linjeskift etter røyropperator.
}
\details{
Lik \code{styler:::add_line_break_after_pipe()},
men der det òg vert lagt til linjeskift etter korte røyr.
}
\keyword{internal}
