% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreger_ki_prop.R
\name{aggreger_ki_prop}
\alias{aggreger_ki_prop}
\title{Regn ut kvalitetsindikator for andeler}
\usage{
aggreger_ki_prop(d_ki_ind, konf_niva = 0.95, alfa = lifecycle::deprecated())
}
\arguments{
\item{d_ki_ind}{Datasett som gitt ut av en standard KI-funksjon
for binomiske data / andelsdata / proporsjonsdata.
Se detaljer nedenfor.}

\item{konf_niva}{Konfidensnivå.
Standardverdi er 0.95, som tilsvarer et 95 \%-konfidensintervall.}

\item{alfa}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Utdatert og erstatta av \code{konf_niva} (tilsvarande \code{1 - alfa}).
Vert fjerna i neste versjon av pakken.}
}
\value{
Ugruppert tibble eller \code{data.frame} (avhengig av inndataene) med følgende
kolonner:
\item{est}{Kvalitetsindikatoren, dvs. estimert andel (\code{ki_teller}/\code{ki_nevner}).}
\item{ki_teller}{Telleren i indikatoren (antall sanne \code{ki_krit_teller}
der \code{ki_krit_nevner} er sann).}
\item{ki_nevner}{Nevneren i indikatoren (antall sanne \code{ki_krit_nevner}).}
\item{konfint_nedre}{Nedre konfidensgrense for \code{est}.}
\item{konfint_ovre}{Øvre konfidensgrense for \code{est}.}
I tillegg vil det være kolonner for alle grupperingsvariablene.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Regner ut kvalitetsindikator for andeler (teller delt på nevner)
basert på et standard datasett for binomiske data / andelsdata.
Gir også ut konfidensintervall for andelene.
}
\details{
Inndatasettet må inneholde minst de to logiske variablene
\code{ki_krit_nevner} og \code{ki_krit_teller}, der \code{ki_krit_nevner}
sier om raden (for eksempel pasienten eller forløpet) oppfyller
kriteriet for å inngå i \emph{nevneren} i indikatoren,
og \code{ki_teller} sier om den oppfyller kriteriet for å inngå i
\emph{telleren} (dvs. skal regnes som en «suksess»). Se vignetten
for kvalitetsindikatorar
(\code{vignette("ki-funksjonar", package = "rapwhale")})
for mer detaljert informasjon, både om krav til gyldige inndata
og om bruk.

Funksjonen gir ut aggregert datasett med teller, nevner, estimatet
teller/nevner og konfidensintervall for estimatet, alt bare for
de radene der \code{ki_krit_nevner} er sann. Dersom ingen \code{ki_krit_nevner}
er sann, vil estimatet og tilhørende konfidensgrenser være \code{NA}.
Hvis inndataene er gruppert, blir resultatet regnet ut på gruppenivå,
med én rad per gruppe.
}
\note{
Funksjonen kan også brukes for teller/nevner-data som ikke oppfyller
antagelsen om binomiske data, men da kan man naturlegvis ikke stole på
konfidensintervallene som en får ut.
}
\examples{
# Pakke for bruk av tibble-objekt og rør-operatoren
library(dplyr)

# Eksempeldata
d = tibble(
  pasid = 1:8,
  sykehus = rep(c("Haukeland", "Førde", "Voss"), times = c(3, 2, 3)),
  ki_krit_teller = c(TRUE, FALSE, TRUE, TRUE, FALSE, TRUE, TRUE, FALSE),
  ki_krit_nevner = c(TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, FALSE)
)

# Utregnet kvalitetsindikator for hele datasettet
aggreger_ki_prop(d)

# Eventuelt med 90 \%-konfidensintervall
aggreger_ki_prop(d, konf_niva = 0.9)

# Gruppert på sykehusnivå
d |>
  group_by(sykehus) |>
  aggreger_ki_prop()

# Merk at sykehusene ovenfor blir vist i alfabetisk rekkefølge,
# siden grupperingsvariabelen «sykehus» var en tekstvariabel.
# Hvis du vil ha en annen rekkefølge, gjør den om til faktor først.
d = mutate(d, sykehus = factor(sykehus,
  levels = c("Haukeland", "Førde", "Voss")
))
d |>
  group_by(sykehus) |>
  aggreger_ki_prop()
}
