% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreger_ki_rate.R
\name{aggreger_ki_rate}
\alias{aggreger_ki_rate}
\title{Regn ut kvalitetsindikator for rater}
\usage{
aggreger_ki_rate(
  d_ki_ind,
  konf_niva = 0.95,
  multiplikator = 1,
  alfa = lifecycle::deprecated()
)
}
\arguments{
\item{d_ki_ind}{Datasett som gitt ut av en standard KI-funksjon for ratedata.
Se detaljer nedenfor.}

\item{konf_niva}{Konfidensnivå.
Standardverdi er 0.95, som tilsvarer et 95 \%-konfidensintervall.}

\item{multiplikator}{Tallverdi som skal multipliseres med raten
(for å vise raten per \code{multiplikator} enheter i utdatene).}

\item{alfa}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Utdatert og erstatta av \code{konf_niva} (tilsvarande \code{1 - alfa}).
Vert fjerna i neste versjon av pakken.}
}
\value{
Ugruppert \code{tibble} eller \code{data.frame}
(avhengig av inndataene)
med følgende kolonner:
\item{est}{Kvalitetsindikatoren, dvs. estimert rate
(sum(\code{ki_antall})/sum(\code{ki_eksponering})*\code{multiplikator}).}
\item{konfint_nedre}{Nedre konfidensgrense for \code{est}.}
\item{konfint_ovre}{Øvre konfidensgrense for \code{est}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Regner ut kvalitetsindikator for rater (antall delt på eksponering),
basert på et standard datasett for ratedata/Poisson-data.
Gir også ut konfidensintervall for ratene.
}
\details{
Inndatasettet må inneholde minst de tre variablene \code{ki_antall},
\code{ki_eksponering} og \code{ki_aktuell},
der \code{ki_antall} er en numeriske variabel som angir hvor mange hendelser
som er blitt observert
(for eksempel antall ganger pasienten har fått smertelindring
eller antall observerte bakteriekolonier),
\code{ki_eksponering} er en numerisk variabel som angir eksponeringen
som var grunnlag for hendelsene i \code{ki_antall}
(her for eksempel antall pasientdøgn
eller størrelsen på undersøkt område, målt i kvadratcentimeter),
og \code{ki_aktuell} er en logisk variabel
som indikerer om observasjonen/raden
skal inkluderes i beregningen eller ikke.

Funksjonen gir ut aggregert datasett med estimert rate og
konfidensintervall for estimatet.
Hvis inndataene er gruppert,
blir resultatet regnet ut på gruppenivå,
med én rad per gruppe.

Om man velger en ikke-standard verdi for \code{multiplikator},
får man ut estimat av rate og tilhørende konfidensintervall
per \code{multiplikator} av enheten til eksponeringen.
Dette kan være hendig om man for eksempel ser på sjeldne hendelser
per innbygger i en stor populasjon.
Da kan man sette \code{multiplikator} til \code{1000} eller \code{100000}
for å få ut raten per 1000 eller 100 000 innbyggere.
}
\note{
Konfidensintervallene blir regnet ut under forutsetning at dataene er
Poisson-fordelt,
og blir da beregnet ved profil-likelihood-metoden.
}
\examples{
# Pakke for bruk av tibble-objekt og rør-operatoren
library(dplyr)

# Eksempeldata
helseforetak = c(
  "Helse Vest",
  "Helse Sør-Øst",
  "Helse Midt",
  "Helse Nord"
)
populasjon = c(1100000, 3100000, 700000, 480000)
antall_behandlinger = c(650000, 1670000, 450000, 350000)
d = tibble(
  helseforetak = helseforetak,
  ki_antall = antall_behandlinger,
  ki_eksponering = populasjon,
  ki_aktuell = c(TRUE, TRUE, FALSE, TRUE)
)

# Utregnet kvalitetsindikator for hele datasettet
aggreger_ki_rate(d)

# Eventuelt med 90 \%-konfidensintervall
aggreger_ki_rate(d, konf_niva = 0.9)

# Gruppert på helseforetak, per 1000
d |>
  group_by(helseforetak) |>
  aggreger_ki_rate(multiplikator = 1000)

# Merk at helseforetakene ovenfor blir vist i alfabetisk rekkefølge,
# siden grupperingsvariabelen «helseforetak» var en tekstvariabel.
# Hvis du vil ha en annen rekkefølge, gjør den om til faktor først.
d = mutate(d, helsefortak = factor(helseforetak,
  levels = c("Helse Nord", "Helse Midt", "Helse Vest", "Helse Sør-Øst")
))
d |>
  group_by(helseforetak) |>
  aggreger_ki_rate()
}
\seealso{
Om du heller er interessert i andelsdata (teller/nevner),
altså data som følger en binomisk fordeling,
bruk \code{\link[=aggreger_ki_prop]{aggreger_ki_prop()}}.
}
