% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreger_ki_snitt.R
\name{aggreger_ki_snitt}
\alias{aggreger_ki_snitt}
\title{Regn ut kvalitetsindikator for gjennomsnitt}
\usage{
aggreger_ki_snitt(d_ki_ind, konf_niva = 0.95, alfa = lifecycle::deprecated())
}
\arguments{
\item{d_ki_ind}{Datasett som gitt ut av en standard KI-funksjon
for numeriske/kontinuerlige data.
Se detaljer nedenfor.}

\item{konf_niva}{Konfidensnivå.
Standardverdi er 0.95, som tilsvarer et 95 \%-konfidensintervall.}

\item{alfa}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Utdatert og erstatta av \code{konf_niva} (tilsvarande \code{1 - alfa}).
Vert fjerna i neste versjon av pakken.}
}
\value{
Ugruppert tibble eller \code{data.frame} (avhengig av inndataene) med følgende
kolonner:
\item{est}{Kvalitetsindikatoren, dvs. gjennomsnittet
av \code{ki_x} for radene som har sanne \code{ki_aktuell}-verdier
(eventuelt per gruppe).}
\item{konfint_nedre}{Nedre konfidensgrense for \code{est}.}
\item{konfint_ovre}{Øvre konfidensgrense for \code{est}.}
\item{n_aktuell}{Antall observasjoner som er inkludert i
beregningen av \code{est} (antall sanne \code{ki_aktuell}).}
I tillegg vil det være kolonner for alle grupperingsvariablene.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Regner ut kvalitetsindikator med gjennomsnitt av kontinuerlige data,
basert på et standard datasett for dette.
Gir også ut konfidensintervall for gjenomsnittet.
}
\details{
Inndatasettet må inneholde minst de to variablene
\code{ki_x} og \code{ki_aktuell}, der \code{ki_x} er den numeriske variabelen
det skal beregnes gjennomsnitt for og \code{ki_aktuell} en
logisk variabel som indikerer om observasjonen skal inkluderes i
beregningen eller ikke.
Se vignetten for kvalitetsindikatorar
(\code{vignette("ki-funksjonar", package = "rapwhale")})
for mer detaljert informasjon, både om krav til gyldige inndata
og om bruk.

Funksjonen gir ut aggregert datasett med estimert gjennomsnitt,
konfidensintervall for estimatet, og antall observasjoner hvor \code{ki_aktuell}
er sann.
Dersom ingen \code{ki_aktuell} er sann eller konfidensintervallet
ikke kan regnes ut (for eksempel fordi det bare finnes én observasjon
i en gitt gruppe), vil konfidensgrensene være \code{NA}.
Hvis inndataene er gruppert, blir resultatet regnet ut på gruppenivå,
med én rad per gruppe.
}
\note{
Konfidensintervallene blir regnet ut under forutsetning
at dataene er normalfordelt.
(De bruker den vanlege formelen basert på \emph{t}-fordelingen.)
Man bør altså være forsiktig med å stole mye på dem
dersom de kommer fra andre fordelinger,
spesielt usymmetriske fordelinger,
med mindre antall observasjoner er så stort
at sentralgrenseteoremet slår inn.
}
\examples{
# Pakke for bruk av tibble-objekt og rør-operatoren
library(dplyr)

# Eksempeldata
d = tibble(
  pasid = 1:8,
  sykehus = rep(c("Haukeland", "Førde", "Voss"), times = c(3, 2, 3)),
  ki_x = c(6.23, 3.05, 12.2, 4.8, 1.2, NA_real_, 1.9, 7.4),
  ki_aktuell = c(TRUE, TRUE, TRUE, TRUE, FALSE, FALSE, TRUE, TRUE)
)

# Utregnet kvalitetsindikator for hele datasettet
aggreger_ki_snitt(d)

# Eventuelt med 90 \%-konfidensintervall
aggreger_ki_snitt(d, konf_niva = 0.9)

# Gruppert på sykehusnivå
d |>
  group_by(sykehus) |>
  aggreger_ki_snitt()

# Merk at sykehusene ovenfor blir vist i alfabetisk rekkefølge,
# siden grupperingsvariabelen «sykehus» var en tekstvariabel.
# Hvis du vil ha en annen rekkefølge, gjør den om til faktor først.
d = mutate(d, sykehus = factor(sykehus,
  levels = c("Haukeland", "Førde", "Voss")
))
d |>
  group_by(sykehus) |>
  aggreger_ki_snitt()
}
