% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapporttekst-funksjoner.R
\name{akse_prosent_format}
\alias{akse_prosent_format}
\title{Formater akse med prosentformat}
\usage{
akse_prosent_format(antall_desimaler = 1, decimal.mark = ",", ...)
}
\arguments{
\item{antall_desimaler}{Antall desimaler skal vises.}

\item{decimal.mark}{Instilling for desimaltall. Standard er ",".}

\item{...}{Ytterligere argumenter.}
}
\value{
En funksjon med spesifiserte innstillinger for prosentformat.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Funksjonen brukes i ggplot-kall og konverterer akselabels til å vise
prosent med riktig format. Standard er mellomrom før prosenttegnet
(slik det skal være på norsk), eks. 0.5 --> "50 \%", og med komma
som desimaltegn. Som standard blir tallene vist
med én desimal, men dette kan endres ved å spesifisere \code{antall_desimaler}.
}
\examples{
# Pakke for å laga figurar
library(ggplot2)

ggplot(iris, aes(x = Sepal.Width / Sepal.Length, y = Petal.Width / Petal.Length)) +
  geom_point() +
  scale_x_continuous(labels = akse_prosent_format()) +
  scale_y_continuous(labels = akse_prosent_format(antall_desimaler = 0))
}
