% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapporttekst-funksjoner.R
\name{akse_tall_format}
\alias{akse_tall_format}
\title{Formater akser med tallformat}
\usage{
akse_tall_format(antall_desimaler = 2, decimal.mark = ",", big.mark = " ", ...)
}
\arguments{
\item{antall_desimaler}{Hvor mange desimaler skal vises.}

\item{decimal.mark}{Instilling for desimaltegn. Standard er ",".}

\item{big.mark}{Instilling for tusenskille. Standard er " ".}

\item{...}{Ytterligere argumenter.}
}
\value{
En funksjon med spesifiserte innstillinger for tallformat.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Funksjonen brukes i ggplot-kall og konverterer akselabels til å vise tall
med tusenskille (standard er " ") og ønsket antall desimaler (standard er 2).
}
\examples{
# Pakke for å laga figurar
library(ggplot2)
library(tibble)

a = tibble(
  x = rnorm(100, 7500, 1000),
  y = runif(100, 0, 1)
)

ggplot(a, aes(x = x, y = y)) +
  geom_point() +
  scale_x_continuous(labels = akse_tall_format(antall_desimaler = 0)) +
  scale_y_continuous(labels = akse_prosent_format(antall_desimaler = 1))
}
