% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema-kvalreg.R
\name{aktiver_kvalregtema}
\alias{aktiver_kvalregtema}
\title{Slå på kvalregtema for ggplot2 og qicharts2}
\usage{
aktiver_kvalregtema()
}
\value{
Returnerer \code{NULL} usynleg.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Set aktivt tema til \code{\link[=tema_kvalreg]{tema_kvalreg()}},
og oppdaterer andre standardverdiar som fargar og punkt-/linjestorleik.
}
\examples{
# Pakke for å laga figurar
library(ggplot2)

# Sølediagram før aktivering av tema
p = ggplot(mpg, aes(class)) +
  geom_bar()
p

aktiver_kvalregtema()

# Søylediagram etter aktivering av tema
p
}
