% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyser_valideringsdatasett.R
\name{analyser_valideringsdatasett}
\alias{analyser_valideringsdatasett}
\title{Analyser valideringsdatasett ved å samanlikna verdiar}
\usage{
analyser_valideringsdatasett(d_vld, samanliknar = samanlikn_identisk)
}
\arguments{
\item{d_vld}{Valideringsdatasett (dataramme/tibble). Sjå
\code{\link[=er_valideringsdatasett_gyldig]{er_valideringsdatasett_gyldig()}} for definisjonen på eit (gyldig)
valideringsdatasett. Det vert automatisk testa at datasettet er gyldig.}

\item{samanliknar}{Samanliknarfunksjon. Standard er \code{\link[=samanlikn_identisk]{samanlikn_identisk()}}.}
}
\value{
Opphavleg datasett, men med to ekstra kolonnar lagd til -
\code{ki_krit_teller} som er \code{TRUE} for kvar rad der verdiane som
skal samanliknast, er «like» i følgje \code{samanliknar}, og \code{FALSE} for
kvar rad der dei ikkje er det, og \code{ki_krit_nevner} som er \code{TRUE} for
alle rader.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Brukar ein samanlikningsfunksjon for å undersøka om interne og eksterne
verdiar i eit valideringsdatasett er «like» eller ikkje.
}
\details{
Funksjonen tek inn eit valideringsdatasett \code{d_vld} og ein samanliknarfunksjon
\code{samanliknar}, og returnerer datasettet med to ekstra kolonnar -
\code{ki_krit_teller} som for kvar rad seier om verdiane som skal
samanliknast, er «like»/ekvivalente (som definert av samanliknarfunksjonen),
og \code{ki_krit_nevner} som er \code{TRUE} for alle rader (denne er med for at utdata
skal vera på formatet \code{aggreger_ki_prop()} nyttar).

For kvar rad i \code{d_vld} der \code{vld_vartype} er for eksempel \code{"x"},
brukar funksjonen \code{samanliknar}-funksjonen til å samanlikna verdiane
i kolonnane \code{vld_verdi_intern_x} og \code{vld_verdi_ekstern_x} (i \code{d_vld}).

Argumentet \code{samanliknar} skal vera ein funksjon \code{f(verdi1, verdi2, varnamn)}
som tek inn tre like lange vektorar (eller siste argument kan vera \code{NULL}).
Dei to fyrste argumenta inneheld verdiane som skal samanliknast elementvis,
og den siste er ein tekstvektor som elementvis seier korleis samanlikninga
skal føregå. I praksis inneheld han namnet på variablane som verdiane i
valideringsdatasettet var henta frå.

Argumentet \code{varnamn} i samanliknarfunksjonen kan vera \code{NULL} dersom han
støttar dette. Det er elles nødvendig å bruka ein samanliknarfunksjon
som forstår verdiane til \code{varnamn}.
Eit eksempel kan vera at ein har \code{varnamn}-element som \code{"vekt_gram"} og
\code{"temperatur"}, der samanliknarfunksjonen veit at at han skal ha større
slingringsmonn når han samanliknar to vektverdiar målt i gram enn to
temperaturmålingar målt i Celsius.
}
\examples{
library(tibble)

# Eksempel på gyldig valideringsdatasett med både «like» og ikkje «like»
# verdiar i følgje standard samanliknar `samanlikn_identisk()`.
d_vld = tibble(
  pasient_id = c(5, 5, 5, 7),
  dato_operasjon = as.Date(c("2020-06-07", "2020-12-13", "2020-12-13", "2021-02-05")),
  kjonn = c("mann", "mann", "mann", "kvinne"),
  sjukehus = c("Bergen", "Bergen", "Førde", "Stavanger"),
  vld_varnamn = c("vekt", "vekt", "diabetes", "diabetes"),
  vld_vartype = c("tal", "tal", "logisk", "logisk"),
  vld_verdi_intern_tal = c(78, 88, NA, NA),
  vld_verdi_ekstern_tal = c(78, 90, NA, NA),
  vld_verdi_intern_logisk = c(NA, NA, TRUE, FALSE),
  vld_verdi_ekstern_logisk = c(NA, NA, TRUE, NA)
)
analyser_valideringsdatasett(d_vld)
}
