% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymiser.R
\name{anonymiser}
\alias{anonymiser}
\title{Anonymiser datavektor}
\usage{
anonymiser(x, startnr = 1001)
}
\arguments{
\item{x}{Vektor (av valfri type/klasse) med verdiar som skal anonymiserast.}

\item{startnr}{Minste verdi for dei tilfeldige tala som skal genererast.}
}
\value{
Heiltalsvektor med anonymiserte verdiar.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Anonymiserer verdiane i ein vektor ved å byta dei ut med tilfeldige tal.
}
\details{
Alle verdiane i  \code{x} vert bytte ut med tilfeldige tal frå følgja
\code{startnr}, \code{startnr + 1}, \code{startnr + 2}, …,
men slik at viss ein verdi finst fleire gongar i \code{x},
får han alltid same (tilfeldige) utverdi.

Innverdiane kan vera av valfri type, for eksempel tekst,
tal eller datoar,
men utverdiane vil alltid vera heiltal.
Eventuelle \code{NA}-verdiar kjem ut som \code{NA}-verdiar,
men med ei åtvaring.
}
\examples{
library(tibble)
library(dplyr)

d = tibble(
  pasient_id = c(2, 5, 7, 5, 8),
  sjukehus = c(
    "Bergen", "Oslo", "Stavanger",
    "Oslo", "Stavanger"
  )
)

mutate(d,
  pasient_id_ano = anonymiser(pasient_id),
  sjukehus_ano = anonymiser(sjukehus, startnr = 11)
)
}
\seealso{
Viss du vil anonymisera verdiar som inngår i fleire datasett,
bruk heller \code{\link[=lag_anonymiseringsfunksjon]{lag_anonymiseringsfunksjon()}}.
}
