% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_eq_linter.R
\name{assignment_eq_linter}
\alias{assignment_eq_linter}
\title{Equals-assignment linter}
\usage{
assignment_eq_linter()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Linter som sjekkar at \code{=} alltid vert brukt ved tilordning.
}
\keyword{internal}
