% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kompletthet.R
\name{beregn_kompletthet}
\alias{beregn_kompletthet}
\title{beregn kompletthet}
\usage{
beregn_kompletthet(data, variabel)
}
\arguments{
\item{data}{Tibble/data.frame som inneholder variabel det skal beregnes
kompletthet for.}

\item{variabel}{Tekststreng med navn på variabel det skal beregnes kompletthet for.}
}
\value{
Returnerer en tibble med følgende kolonner: \cr
grupperingsvariabel - Hvis inndata er gruppert vil grupperingsvariabel være
med i utdata. \cr
Variabel - Navn på variabel det er beregnet kompletthet for. \cr
Totalt_antall - Totalt antall observasjoner per gruppe i inndata  \cr.
Antall_na - Antall observasjoner som er NA i inndata. \cr
Andel_na - Andel observasjoner som er NA i inndata.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Regner ut antall og andel manglende observasjoner for en variabel. Tar inn
et datasett og en variabel det skal beregnes kompletthet for.
}
\examples{
# Pakke for bruk av tibble-objekt og rør-operatoren
library(dplyr)

# Ugrupperte inndata:
d = tibble(
  pas_id = c(1, 2, 3, 4, 5, 6),
  sykehus = c("HUS", "HUS", "SVG", "SVG", "SVG", "OUS"),
  var_1 = c(1, 2, -1, 99, NA, 5)
)
beregn_kompletthet(data = d, variabel = "var_1")

# Grupperte inndata:
d = tibble(
  pas_id = c(1, 2, 3, 4, 5, 6),
  sykehus = c("HUS", "HUS", "SVG", "SVG", "SVG", "OUS"),
  var_1 = c(1, 2, -1, 99, NA, 5)
) |>
  group_by(sykehus)

beregn_kompletthet(data = d, variabel = "var_1")
}
