% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kompletthet.R
\name{beregn_kompletthet_datasett}
\alias{beregn_kompletthet_datasett}
\title{Beregn kompletthet for datasett}
\usage{
beregn_kompletthet_datasett(data)
}
\arguments{
\item{data}{Datasett det skal beregnes kompletthet for.}
}
\value{
Returnerer en tibble med oversikt over antall observasjoner, antall NA og
andel NA for hver variabel.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Tar inn et datasett det skal beregnes kompletthet for og returnerer en
tibble med oversikt over kompletthet for samtlige variabler. Tar ikke
hensyn til Ukjent/ikke besvart etc.
}
\examples{
library(tibble)

# Datasett det skal beregnes kompletthet for
d_eksempel = tibble(
  pas_id = c(1L, 2L, 3L, 4L, 5L, 6L),
  sykehus = c("HUS", "HUS", "SUS", "SUS", "SUS", "OUS"),
  vekt = c(60L, NA_integer_, 100L, NA_integer_, 99L, -1L),
  vekt_2 = c(55L, NA_integer_, 99L, -1L, NA_integer_, 50L),
  hoyde = c(1.52, NA_real_, 1.89, 2.15, NA_real_, 99.9),
  symptom = c("svett", "klam", NA_character_, "trøtt", "vet ikke", "Ukjent"),
  test_logisk = c(TRUE, FALSE, NA, NA, FALSE, TRUE)
)

beregn_kompletthet_datasett(data = d_eksempel)
}
