% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kompletthet.R
\name{beregn_kompletthet_med_ukjent}
\alias{beregn_kompletthet_med_ukjent}
\title{Beregn kompletthet med ukjent}
\usage{
beregn_kompletthet_med_ukjent(data, variabel, na_vektor)
}
\arguments{
\item{data}{Tibble/data.frame som inneholder variabel det skal beregnes
kompletthet for.}

\item{variabel}{Streng med navn p책 variabel det skal beregnes kompletthet for.}

\item{na_vektor}{Vektor som angir hvilke verdier i 'variabel' som skal
erstattes med NA.}
}
\value{
Returnerer et aggregert datasett med kolonnene: \cr
Grupperingsvariabel - hvis inndata er gruppert.\cr
Variabel - Navn for variabel det er beregnet kompletthet for.\cr
Totalt_antall - Antall rader per gruppe.\cr
Antall_na - Antall NA i opprinnelig data.\cr
Andel_na - Andel NA i opprinnelig data.\cr
Antall_na_med_ukjent - Antall NA n책r verdier fra NA-vektor er erstattet med NA.\cr
Andel_na_med_ukjent - Andel NA n책r verdier er NA-vektor er erstattet med NA.\cr
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Beregner antall og andel NA, b책de med og uten ukjente verdier.
Ukjente verdier defineres i na_vektor. Tanken er at besvarelser som
"Velg Verdi", "Ukjent", "Ikke besvart" eller lignende kan inkluderes her.
}
\examples{
library(tibble)

d = tibble(
  pas_id = c(1, 2, 3, 4, 5, 6),
  sykehus = c("A", "A", "B", "B", "B", "C"),
  var_1 = c(1, 2, -1, 99, NA, 5)
)
beregn_kompletthet_med_ukjent(data = d, variabel = "var_1", na_vektor = c(-1, 99))
}
