% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boy_enhet.R
\name{boy_enhet}
\alias{boy_enhet}
\title{Formater tall med automatisk entalls-/flertallsenhet}
\usage{
boy_enhet(x, entall, flertall, formatering = erstatt_0, ...)
}
\arguments{
\item{x}{Numerisk vektor med minst ett element.
Kan ikke inneholde \code{NA}-verdier.}

\item{entall}{Tekst som skal brukes som enhet for \code{x}-elementer lik \code{1}
eller \code{-1}.}

\item{flertall}{Tekst som skal brukes som enhet for \code{x}-elementer ulik \code{1}
og \code{-1}.}

\item{formatering}{Funksjon for formatering av verdiene i \code{x}.
Tar inn \code{x} og må gi ut en vektor av samme lengde,
enten en tekstvektor eller en vektor som kan
gjøres om til tekst via \code{\link[=as.character]{as.character()}}.}

\item{...}{Ytterligere argumenter sendt til \code{formatering()}.}
}
\value{
Tekstvektor av samme lengde som \code{x}.
Hver tallverdi i \code{x} blir formatert som tallverdien + mellomrom + enhet.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Formaterer tall med tilhørende enhet automatisk bøyd til
entalls-/flertallsform.
}
\details{
Formaterer en tallvektor,
\code{x},
med tilhørende enhet bøyd til
enten entalls- eller flertallsform.
For eksempel kan tallet 1 bli formatert som «1 pasient»,
mens tallet 34 blir formatert som «34 pasienter».

Alle \code{x}-elementer lik 1 eller -1 blir formatert med enheten \code{entall},
mens resterende elementer blir formatert med enheten \code{flertall}.

Formateringsfunksjonen \code{formatering} kan for eksempel
brukes til å formatere tallene i \code{x} med tusenskilletegn
eller avrundet til heltal.
Standard formateringsfunksjon \code{\link[=erstatt_0]{erstatt_0()}}
gjør i utgangspunktet \emph{ingen} formatering,
men tar argumentet \code{nullverdi},
som kan brukes til å erstatte tallet 0
med en selvvalgt verdi,
for eksempel teksten «ingen»
(se eksempel nedenfor).
}
\note{
Funksjonen håndterer også desimaltall,
med samme regler,
også for tall der desimaldelen slutter på 1.
Det blir for eksempel «2,1 millioner»,
ikke «2,1 million»,
selv om begge skrivemåtene er tillatt på norsk.
}
\examples{
# Formatering av enkeltverdier
boy_enhet(1, "operasjon", "operasjoner")
boy_enhet(34, "operasjon", "operasjoner")

# Formatering av vektorverdier, og med ikke-standard nullverdi
boy_enhet(c(0, 1, 2),
  entall = "ny pasient",
  flertall = "nye pasienter",
  nullverdi = "ingen"
)

# Formatering til LaTeX-tekst og med ett desimaltall
boy_enhet(c(2.67, 1, 0, pi),
  entall = "million",
  flertall = "millioner",
  formatering = num,
  desimalar = 1
)
}
