% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bygg_og_installer_pakker.R
\name{bygg_og_installer_pakker}
\alias{bygg_og_installer_pakker}
\title{Bygg og installer R-pakker}
\usage{
bygg_og_installer_pakker(pakkemapper, installer = TRUE, stille = TRUE)
}
\arguments{
\item{pakkemapper}{Tekstvektor med mappeadresser til R-pakker.}

\item{installer}{Logisk variabel. Skal pakkene installeres (som standard ja).}

\item{stille}{Logisk variabel. Skal det kun gis ut oppsummerte advarsler og
feilmeldinger (som standard ja)?}
}
\value{
\code{NULL}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Dokumenterer, bygger og installerer R-pakker.
}
\details{
For hver mappeadresse i \code{pakkemapper} bygges først pakkedokumentasjonen
og så pakken (som en \code{.tar.gz}-pakkefil).
Hvis \code{installer = TRUE}, blir også pakken installert.
Hvis \code{stille = TRUE},
blir det kun gitt ut oppsummerte advarsler og feilmeldinger.
}
\examples{
\dontrun{
# Finn først alle DESCRIPTION-filer, som indikerer
# at tilhørende foreldermappe er en R-pakkemappe
desc_filer = list.files(
  "m:\\\\r-pakker\\\\",
  pattern = "^DESCRIPTION$",
  full.names = TRUE, recursive = TRUE
)

# Hent ut mappeadressene, og bygg og installer pakkene
pakkemapper = dirname(desc_filer)
bygg_og_installer_pakker(pakkemapper)
}
}
