% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadump-valider.R
\name{dd_er_gyldig}
\alias{dd_er_gyldig}
\title{Validering av datadump}
\usage{
dd_er_gyldig(d, kodebok, ...)
}
\arguments{
\item{d}{Datasett som skal valideres.}

\item{kodebok}{Kodebok med informasjon om variablene i valgt datadump.
Skal være kodebok på kanonisk form og på Nasjonalt servicemiljø for
medisinske kvalitetsregistre region vest (NASERVE)
sitt standard kodebokformat.}

\item{...}{Eventuelle argument som blir gitt videre til \code{\link[=lag_regelsett]{lag_regelsett()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Funksjon som automatisk lager et regelsett basert på en kodebok, og sjekker at
datadumpen er gyldig gitt dette regelsettet. \cr \cr
Returnerer TRUE viss og berre viss datadumpen er gyldig. \cr
Hvis FALSE, vil funksjonen gi ut en oppsummering på et format fra ruler-pakken
på hva som er feil og hvor disse er å finne.
}
