% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logiske-tester.R
\name{ekviv}
\alias{ekviv}
\alias{\%ekviv\%}
\title{Sjekk om sann \code{a} er ekvivalent med sann \code{b}}
\usage{
ekviv(a, b)

a \%ekviv\% b
}
\arguments{
\item{a}{Logisk vektor som elementvis skal sjekkes mot \code{b}.}

\item{b}{Logisk vektor som elementvis skal sjekkes mot \code{a}.}
}
\value{
Logisk vektor som elementvis sier om \code{a = TRUE} er ekvivalent
med \code{b = TRUE}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Tar inn to logiske vektorer \code{a} og \code{b}, og gir ut en logisk vektor som
elementvis sier om \code{a = TRUE} er ekvivalent med \code{b = TRUE}.
}
\details{
Funksjonen tar inn to logiske vektorer \code{a} og \code{b}, sjekker elementvis om
\code{a = TRUE} er ekvivalent med \code{b = TRUE}.

Den gir ut en logisk vektor som er \code{TRUE} for hvert elementpar fra
\code{a} og \code{b} der både \code{a} og \code{b = TRUE}, eller både \code{a} og \code{b = FALSE}
eller \code{NA}, og \code{FALSE} ellers. Den gir altså ut \code{FALSE} for
elementpar som motstrider at \code{a = TRUE} er ekvivalent med \code{b = TRUE}.

Funksjonen tilsvarer altså \code{a == b}, men håndterer \code{NA} som \code{FALSE}.

Hvis \code{a} og \code{b} har ulik lengde blir den korteste brukt på ny nødvendig
antall ganger, likt som for logiske operatorer. Se \link[base:Logic]{base::Logic} for
detaljer.
}
\section{Functions}{
\itemize{
\item \code{a \%ekviv\% b}: \code{\link[=ekviv]{ekviv()}} som infiks-operator

}}
\examples{
library(tibble)

d = tribble(
  ~pas_id, ~operert, ~komplikasjoner,
  1, "Ja", "Nei",
  2, "Ja", "Nei",
  3, "Nei", NA,
  4, "Nei", NA,
  5, "Ja", "Ja"
)

er_operert = d$operert == "Ja"
komplikasjoner_registrert = !is.na(d$komplikasjoner)

# Sjekk om alle som er operert har ein verdi
# registrert i variabelen komplikasjoner
ekviv(er_operert, komplikasjoner_registrert)

# Eller som infiks-operator
er_operert \%ekviv\% komplikasjoner_registrert
}
