% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fodselsnummer-funksjoner.R
\name{er_fnr_sjekksum_korrekt}
\alias{er_fnr_sjekksum_korrekt}
\title{Kontroller sjekksum for fødselsnummer og liknande}
\usage{
er_fnr_sjekksum_korrekt(nummer)
}
\arguments{
\item{nummer}{Tekstvektor der kvart element har 11 siffer.}
}
\value{
Logisk vektor som gjev ut \code{TRUE} eller \code{FALSE} for kvart element
i \code{nummer} alt etter om det høvesvis gjev korrekt sjekksum eller ikkje.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Tek inn ein tekstvektor der kvart element har 11 siffer,
og gjev for kvart element ut \code{TRUE} viss dei to siste siffera,
kontrollsiffera, er korrekt i høve dei føregåande siffera,
og \code{FALSE} elles.
}
\details{
Kvart av kontrollsiffera skal vera ein sjekksum rekna ut frå dei
føregåande siffera. Viss dei ni fyrste siffera av eit fødselsnummer,
eller anna liknande nummer, er d1 d2 m1 m2 å1 å2 i1 i2 i3,
skal kontrollsiffera, k1 og k2, oppfylla fylgjande:
\itemize{
\item \verb{k1 = 11 - ((3 * d1 + 7 * d2 + 6 * m1 + 1 * m2 + 8 * å1 + 9 * å2 + 4 * i1 + 5 * i2 + 2 * i3) mod 11)}
\item \verb{k2 = 11 - ((5 * d1 + 4 * d2 + 3 * m1 + 2 * m2 + 7 * å1 + 6 * å2 + 5 * i1 + 4 * i2 + 3 * i3 + 2 * k1) mod 11)}.
}
}
\examples{
nummer = c(
  "15076500565", # Gyldig F-nummer
  "15076500575"
) # Feil i kontrollsiffer

rapwhale:::er_fnr_sjekksum_korrekt(nummer)
}
\keyword{internal}
