% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fodselsnummer-funksjoner.R
\name{er_gyldig_f_nummer}
\alias{er_gyldig_f_nummer}
\title{Sjekk gyldig F-nummer}
\usage{
er_gyldig_f_nummer(nummer)
}
\arguments{
\item{nummer}{Tekstvektor med nummer som skal sjekkast.}
}
\value{
Logisk vektor som gjev ut \code{TRUE} eller \code{FALSE} for kvart element
i \code{nummer} alt etter om det høvesvis er eit gyldig F-nummer eller ikkje.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Tek inn ein tekstvektor,
og gjev for kvart element ut \code{TRUE} viss det er eit gyldig fødselsnummer,
og \code{FALSE} elles.
}
\details{
I eit gyldig fødselsnummer er dei fyrste seks siffera ein fødselsdato på
formatet "DDMMYY", og dei siste fem siffera eit personnummmer.
Personnummeret inneheld tre individsiffer, og to kontrollsiffer rekna ut
frå dei føregåande siffera.
}
\examples{
nummer = c(
  "15076500565", # Gyldig F-nummer
  "70019950032", # Gyldig D-nummer
  "01410199935", # Gyldig H-nummer
  "88888888831", # Gyldig FH-nummer
  "98019800546" # Ugyldig nummer generelt
)

rapwhale:::er_gyldig_f_nummer(nummer)
}
\keyword{internal}
