% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fodselsnummer-funksjoner.R
\name{er_gyldig_fh_nummer}
\alias{er_gyldig_fh_nummer}
\title{Sjekk gyldig FH-nummer}
\usage{
er_gyldig_fh_nummer(nummer)
}
\arguments{
\item{nummer}{Tekstvektor med nummer som skal sjekkast.}
}
\value{
Logisk vektor som gjev ut \code{TRUE} eller \code{FALSE} for kvart element
i \code{nummer} alt etter om det høvesvis er eit gyldig FH-nummer eller ikkje.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Tek inn ein tekstvektor,
og gjev for kvart element ut \code{TRUE} viss det er eit gyldig FH-nummer,
og \code{FALSE} elles.
}
\details{
Eit gyldig FH-nummer er eit ellevesifra nummer
der fyrste siffer er 8 eller 9,
dei neste åtte siffera er tilfeldige,
og dei to siste siffera er kontrollsiffer rekna ut frå dei føregåande siffera.
}
\examples{
nummer = c(
  "15076500565", # Gyldig F-nummer
  "70019950032", # Gyldig D-nummer
  "01410199935", # Gyldig H-nummer
  "88888888831", # Gyldig FH-nummer
  "98019800546" # Ugyldig nummer generelt
)

rapwhale:::er_gyldig_fh_nummer(nummer)
}
\keyword{internal}
