% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fodselsnummer-funksjoner.R
\name{er_gyldig_fnr_dato}
\alias{er_gyldig_fnr_dato}
\title{Sjekk gyldig dato}
\usage{
er_gyldig_fnr_dato(dato)
}
\arguments{
\item{dato}{Tekstvektor der kvart element har 6 siffer.}
}
\value{
Logisk vektor som gjev ut \code{TRUE} eller \code{FALSE} for kvart element
i \code{dato} alt etter om det høvesvis er ein gyldig dato eller ikkje.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Tek inn ein tekstvektor med sekssifra element og gjev for kvart element ut
\code{TRUE} viss det er ein gyldig dato på formatet "DDMMYY", og \code{FALSE} elles.
}
\examples{
datoar = c(
  "010101",
  "290204", # Skotårsdag
  "290225", # «Skotårsdag» som aldri er gyldig ...
  "320101"
)

rapwhale:::er_gyldig_fnr_dato(datoar)
}
\keyword{internal}
