% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fodselsnummer-funksjoner.R
\name{er_syntaktisk_fnr}
\alias{er_syntaktisk_fnr}
\title{Sjekk syntaktisk fødselsnummer}
\usage{
er_syntaktisk_fnr(nummer)
}
\arguments{
\item{nummer}{Tekstvektor.}
}
\value{
Logisk vektor som gjev ut \code{TRUE} eller \code{FALSE} for kvart element
i \code{nummer} alt etter om det høvesvis er syntaktisk korrekt eller ikkje.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Tek inn ein tekstvektor og gjev for kvart element ut \code{TRUE} viss det er
11 teikn langt og berre inneheld siffer, og \code{FALSE} elles.
}
\examples{
nummer = c("12345612345", "123456789", "123456789ab", "abcdefghijk")

rapwhale:::er_syntaktisk_fnr(nummer)
}
\keyword{internal}
