% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_valideringsdatasett_gyldig.R
\name{er_valideringsdatasett_gyldig}
\alias{er_valideringsdatasett_gyldig}
\title{Valider valideringsdatasett}
\usage{
er_valideringsdatasett_gyldig(d_vld)
}
\arguments{
\item{d_vld}{Valideringsdatasett (dataramme/tibble).}
}
\value{
\code{TRUE} dersom datasettet er gyldig og \code{FALSE} dersom det
ikkje er det.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Tek inn eit valideringsdatasett og returnerer \code{TRUE} eller \code{FALSE} dersom
det er høvesvis gyldig eller ugyldig.
}
\details{
Eit gyldig valideringsdatasett består av éin eller fleire
indeksvariablar, variablane \code{vld_varnamn} og \code{vld_vartype} og
to eller fleire verdivariablar.

Eksempel på indeksvariablar er pasient- og/eller forløps-ID pluss ekstra
metadata som kjønn, sjukehus og operasjonstype. Saman med \code{vld_varnamn}
skal indeksvariablane unikt identifisera ei rad (observasjonen som skal
validerast).

Variabelen \code{vld_varnamn} inneheld namnet på variabelen som rada
inneheld valideringsdata for, mens \code{vld_vartype} seier kva
type variabel det er, for eksempel tal, tekst eller dato.
Ein kan fritt velja kva verdiar ein brukar for \code{vld_vartype}
(for eksempel \code{"tal"}, \code{"tall"}, \code{"heiltal"} eller \code{"int32"}),
men verdien må starta med ein bokstav og berre innehalda
bokstavar, understrek og/eller siffer.

For kvar unike \code{vld_vartype} skal det finnast verdivariablar –
variablar med prefiksa \code{vld_verdi_intern_} og \code{vld_verdi_ekstern_}.
Viss for eksempel ein \code{vld_vartype} er \code{"tal"}, skal det finnast
verdivariablar \code{vld_verdi_intern_tal} og \code{vld_verdi_ekstern_tal}.
Den første seier kva verdi variabelen i \code{vld_varnamn} har i den
\emph{interne} kjelda, som typisk er eit kvalitetsregister, og den
andre seier kva verdi han har i den \emph{eksterne} kjelda, som typisk
er ein pasientjournal eller eit anna fagsystem. Dei to variablane
skal naturlegvis ha tilsvarande variabeltype som \code{vld_vartype},
for \code{"tal"} for eksempel' \code{double} eller \code{integer}.

Det skal ikkje finnast andre kolonnar som startar med \code{vld_} enn
dei som er nemnde ovanfor (\code{vld_} er eit reservert prefiks).
}
\examples{
library(tibble)

# Eksempel på gyldig valideringsdatasett med to ulike variabeltypar,
# "tal" («double») og "logisk" («logical»). Merk at variabelen «vekt»
# har to verdiar for same pasient, men det er OK, sidan det var på to
# ulike operasjonsdatoar.
d_vld_gyldig = tibble(
  pasient_id = c(5, 5, 5, 7),
  dato_operasjon = as.Date(c("2020-06-07", "2020-12-13", "2020-12-13", "2021-02-05")),
  kjonn = c("mann", "mann", "mann", "kvinne"),
  sjukehus = c("Bergen", "Bergen", "Førde", "Stavanger"),
  vld_varnamn = c("vekt", "vekt", "diabetes", "diabetes"),
  vld_vartype = c("tal", "tal", "logisk", "logisk"),
  vld_verdi_intern_tal = c(78, 88, NA, NA),
  vld_verdi_ekstern_tal = c(78, 90, NA, NA),
  vld_verdi_intern_logisk = c(NA, NA, TRUE, FALSE),
  vld_verdi_ekstern_logisk = c(NA, NA, TRUE, NA)
)

er_valideringsdatasett_gyldig(d_vld_gyldig) # TRUE

# Eksempel på ugyldig valideringsdatasett, der indeksvariablane
# pluss «vld_varnamn» ikkje unikt identifiserer ei rad.
d_vld_ugyldig = d_vld_gyldig
d_vld_ugyldig$dato_operasjon[2] = d_vld_ugyldig$dato_operasjon[1]

er_valideringsdatasett_gyldig(d_vld_ugyldig) # FALSE
}
