% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erstatt_0.R
\name{erstatt_0}
\alias{erstatt_0}
\title{Erstatt tall lik 0 med valgfri verdi}
\usage{
erstatt_0(x, nullverdi = 0)
}
\arguments{
\item{x}{Numerisk vektor.}

\item{nullverdi}{Verdi (typisk tall eller tekst)
som skal erstatte \code{x}-elementer lik 0.}
}
\value{
Vektor av samme lengde og innhold som \code{x},
med 0-verdier byttet ut med \code{nullverdi}.
Verdiene blir om nødvendig gjort om til
å ha same type som \code{nullverdi}
(oftest tekst eller tall),
etter standard R-regler.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Erstatter alle tall som er lik 0 med en valgfri verdi.
}
\details{
Alle verdier nøyaktig lik 0
blir erstattet med verdien \code{nullverdi}.
Funksjonen kan for eksempel brukes til
å erstatte tallet 0 med teksten «ingen» i løpende tekst.

Funksjonen er hovedsakelig laget for å brukes
sammen med \code{\link[=boy_enhet]{boy_enhet()}}.
}
\examples{
erstatt_0(c(0, 1, 2), "ingen")
erstatt_0(c(0, 1, 2), -99)
erstatt_0(c(0, 1, 2), NA)

# Merk at utdata bare blir gjort om til tekst *om nødvendig*
erstatt_0(c(2, 1), "ingen") # Tall
erstatt_0(c(2, 0, 1), "ingen") # Tekst
}
