% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kompletthet.R
\name{erstatt_ukjent}
\alias{erstatt_ukjent}
\title{Erstatt ukjent verdi}
\usage{
erstatt_ukjent(data, variabel, na_vektor)
}
\arguments{
\item{data}{tibble/data.frame som inneholder variabel hvor
ukjente verdier skal erstattes.}

\item{variabel}{tekststreng med navn på variabel hvor ukjente verdier skal
erstattes.}

\item{na_vektor}{vektor med verdier som skal erstattes med NA.}
}
\value{
Returnerer opprinnelig datasett hvor verdier i 'na_vektor' funnet i
'variabel' er erstattet med NA.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Erstatter utvalgte 'ukjente' verdier med NA. Noen datasett har variabler
hvor manglende besvarelser får verdi -1, 99 eller andre verdier for å
indikere at spørsmålet ikke er besvart. I tillegg kan det også finnes
ekte NA-verdier, slik at beregning av kompletthet kompliseres.
Funksjonen tar inn en na_vektor som indikerer hvilke verdier som
skal erstattes med NA, slik at beregning av NA kan inkludere
disse "ukjente" besvarelsene.
}
\examples{
library(tibble)

d = tibble(
  pas_id = c(1, 2, 3, 4, 5, 6),
  var_1 = c(1, 2, -1, 99, NA, 5)
)
erstatt_ukjent(data = d, variabel = "var_1", na_vektor = c(-1, 99))
}
