% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kompletthet.R
\name{erstatt_ukjent_for_datasett}
\alias{erstatt_ukjent_for_datasett}
\title{Erstatt ukjent for datasett}
\usage{
erstatt_ukjent_for_datasett(data, ukjent_datasett)
}
\arguments{
\item{data}{Tibble/data.frame som inneholder variablene det skal
erstattes ukjente verdier med NA for.}

\item{ukjent_datasett}{Tibble med variabelnavn og oversikt over hvilke
verdier som skal erstattes med NA. Se eksempel for hvilken struktur
ukjent_datasett må ha.}
}
\value{
Returnerer opprinnelig datasett men verdier oppgitt i \emph{ukjent_datasett} er
erstattet med NA for de aktuelle variablene.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Enkelte variabler har egne verdier for å indikere manglende besvarelse
i tillegg til NA-verdier, for eksempel verdien -1 for
verditekst "Velg Verdi" eller "Ukjent verdi".
Funksjonen erstatter disse verdiene med NA for å beregne andel manglende
besvarelser inkludert disse ukjente verdiene.
Blir brukt av funksjonen \code{beregn_kompletthet_datasett_med_ukjent()}.
}
\examples{
library(tibble)

d_eksempel = tibble(
  pas_id = c(1L, 2L, 3L, 4L, 5L, 6L),
  sykehus = c("HUS", "HUS", "SUS", "SUS", "SUS", "OUS"),
  vekt = c(60L, NA_integer_, 100L, NA_integer_, 99L, -1L),
  vekt_2 = c(55L, NA_integer_, 99L, -1L, NA_integer_, 50L),
  hoyde = c(1.52, NA_real_, 1.89, 2.15, NA_real_, 99.9),
  symptom = c("svett", "klam", NA_character_, "trøtt", "vet ikke", "Ukjent"),
  test_logisk = c(TRUE, FALSE, NA, NA, FALSE, TRUE)
)

ukjent_datasett_eksempel = tibble(
  variabel = c(
    "pas_id",
    "sykehus", rep("vekt", 2),
    rep("vekt_2", 2),
    "hoyde", rep("symptom", 2),
    "test_logisk"
  ),
  ukjent_verdi_integer = c(
    NA_integer_, NA_integer_, 99,
    -1, 99, -1, NA_integer_, NA_integer_, NA_integer_, NA_integer_
  ),
  ukjent_verdi_real = c(
    NA_real_, NA_real_, NA_real_, NA_real_,
    NA_real_, NA_real_, 99.9, NA_real_, NA_real_, NA_real_
  ),
  ukjent_verdi_tekst = c(
    NA_character_, NA_character_,
    NA_character_, NA_character_, NA_character_, NA_character_,
    NA_character_, "vet ikke", "Ukjent", NA_character_
  )
)

erstatt_ukjent_for_datasett(
  data = d_eksempel,
  ukjent_datasett = ukjent_datasett_eksempel
)
}
