% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema-kvalreg.R
\name{expand_soyle}
\alias{expand_soyle}
\alias{expand_soyle_str_fig}
\title{Juster søylediagram}
\usage{
expand_soyle()

expand_soyle_str_fig()
}
\value{
Returnerer en tall-vektor som brukes som expand-argument i
scale_* funksjoner.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Disse funksjonene brukes for å justere plassering av søyler i søylediagram.
Søylene starter helt inntil aksen og får litt luft i enden av søylen.
}
\details{
\itemize{
\item \code{expand_soyle()}: Brukes for å flytte start av søyle helt
inn til aksen og legger til litt luft i enden av søylen.
\item \code{expand_soyle_str_fig()}: Brukes i de tilfeller der det ikke
er tilstrekkelig plass til tekst-label i plot. Gir litt større
plass i enden av søylene.
}
}
\examples{
# Pakke for å laga figurar
library(ggplot2)

p = ggplot(mtcars, aes(x = cyl, y = mpg)) +
  geom_col()

# Default plot har ikke start av søyle helt inntil aksen
p
p + scale_y_continuous(expand = expand_soyle())

# Legge til mer rom i enden av søylene for tekst-labels
p = ggplot(mtcars, aes(x = cyl)) +
  geom_bar() +
  geom_text(stat = "count", aes(x = cyl, label = ..count..), vjust = -1)

p
p + scale_y_continuous(expand = expand_soyle_str_fig())
}
