% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farger-kvalreg.R
\name{farge_morkare}
\alias{farge_morkare}
\title{Lag mørkare/lysare fargar}
\usage{
farge_morkare(fargar, grad = 5)
}
\arguments{
\item{fargar}{ein vektor med fargar.}

\item{grad}{seier kor mykje mørkare fargen skal gjerast (so bruk negative verdiar for å gjera han lysare).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Gjer ein vektor med fargar mørkare.
Brukar CIELAB-fargerommet til utrekningar
(i staden for RGB-fargerommet), for
betre resultat (meir tilpassa korleis
synet vårt fungerer).
}
\examples{
# Vektor med fargar
fargar = c("#000059", "#084594", "#2171b5")

# Lag vektor med fargar mørkare
fargar_morkare = farge_morkare(fargar)
fargar_morkare

# Lag vektor med fargar lysare
fargar_lysare = farge_morkare(fargar, grad = -5)
fargar_lysare

scales::show_col(c(fargar_morkare, fargar, fargar_lysare))
}
