% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farger-kvalreg.R
\name{farger_kvalreg}
\alias{farger_kvalreg}
\title{Lagre offisielle fargekoder i miljøet}
\usage{
farger_kvalreg()
}
\value{
En liste med \code{farger_hoved}, \code{farger_noyt} og \code{farger_kontr}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Funksjonen gir ut en statisk liste med fargekoder for de offisielle fargene
som skal brukes i figurer i årsrapporter.
}
\examples{
# les inn fargekoder
farger = farger_kvalreg()

farger

scales::show_col(unlist(farger))
}
