% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fodselsnummer-funksjoner.R
\name{finn_type_idnummer}
\alias{finn_type_idnummer}
\title{Finn type ID-nummer}
\usage{
finn_type_idnummer(nummer)
}
\arguments{
\item{nummer}{Tekstvektor med nummer som skal sjekkast.}
}
\value{
Tekstvektor som elemetvis seier kva type ID-nummer elementa
i \code{nummer} er.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Tek inn vektor med \code{nummer} som skal sjekkast,
og gjev ut ein tekstvektor som elementvis seier kva type ID-nummer det er,
og \code{NA} for element som ikkje er ein av dei støtta typane ID-nummer
(fødselsnummer, D-nummer, H-nummer og FH-nummer). Sjå \code{\link[=fnr_er_gyldig]{fnr_er_gyldig()}}
for meir info om dei ulike typane.
}
\examples{
nummer = c(
  "15076500565", # Gyldig F-nummer
  "70019950032", # Gyldig D-nummer
  "01410199935", # Gyldig H-nummer
  "88888888831", # Gyldig FH-nummer
  "98019800546" # Ugyldig nummer generelt
)

finn_type_idnummer(nummer)
}
