% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regn-sumskaar-generelt.R
\name{finn_ugyldige_verdier}
\alias{finn_ugyldige_verdier}
\title{Finn ugyldige verdier i et datasett}
\usage{
finn_ugyldige_verdier(d, verditabell)
}
\arguments{
\item{d}{Dataramme/tibble som kun inneholder kolonner med identiske
navn som i \code{verditabell$variabel}. Alle kolonnene må inneholde
numeriske verdier.}

\item{verditabell}{Dataramme/tibble med to kolonner,
\code{variabel} og \code{verdi},
som sier hvilke verdier som er gyldige for hvilke
variabler. Kolonnen \code{variabel} må være av typen tekst og
\code{verdi} må være numerisk.}
}
\value{
Tibble med én rad for hver ugyldige verdi i \code{d}.
Inneholder tre kolonner, \code{radnr}, \code{variabel} og \code{feilverdi},
som tilsvarer radnummer, kolonnenavn og celleverdi i \code{d}.
Radene er sortert etter radnummer og så kolonnerekkefølge i \code{d}.
Hvis \code{d} ikke inneholder noen ugyldige verdier, vil tibble-en ha
null rader.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Gir ut en tibble med oversikt over alle ugyldige verdier i datasettet,
basert på en oversikt over mulige gyldige verdier.
}
\details{
Blir brukt av \code{\link[=sjekk_variabelverdier]{sjekk_variabelverdier()}}, men kan også
brukes separat. Resultatene kan oppsummeres til en
tekststreng med \code{\link[=oppsummer_ugyldige_verdier]{oppsummer_ugyldige_verdier()}}.
}
\note{
Verdien \code{NA} blir \emph{ikke} håndtert på noen spesiell måte.
Hvis denne skal regnes som gyldig,
må den altså eksplisitt listes opp i \code{verditabell}
(for hver \code{variabel}).
}
\examples{
# To ugyldige verdier
d = data.frame(kjonn = c(2, 1, 1, 3), livskvalitet = c(0, 10, NA, 5))
verditabell = data.frame(
  variabel = c(rep("kjonn", 2), rep("livskvalitet", 3)),
  verdi = c(1, 2, 0, 5, 10)
)
rapwhale:::finn_ugyldige_verdier(d, verditabell)

# Berre gyldige verdier
rapwhale:::finn_ugyldige_verdier(d[1:2, ], verditabell)
}
\seealso{
\code{\link[=sjekk_variabelverdier]{sjekk_variabelverdier()}}, \code{\link[=oppsummer_ugyldige_verdier]{oppsummer_ugyldige_verdier()}}
}
\keyword{internal}
