% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema-kvalreg.R
\name{fjern_x}
\alias{fjern_x}
\alias{fjern_y}
\alias{fjern_x_ticks}
\alias{fjern_y_ticks}
\title{Temainstillinger for figurer}
\usage{
fjern_x()

fjern_y()

fjern_x_ticks()

fjern_y_ticks()
}
\value{
Funksjonene returnerer et theme-objekt som kan brukes direkte i et ggplot-kall.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Disse funksjonene brukes for å gjøre mindre endringer for panellinjer og
strekmarkeringer på aksene i figurer.
De krever ingen argumenter og legges til direkte i et ggplot-kall.
}
\details{
Funksjoner som er inkludert her:
\itemize{
\item \code{fjern_x()}: Fjerner panellinjer på x-aksen.
\item \code{fjern_y()}: Fjerner panellinjer på y-aksen.
\item \code{fjern_x_ticks()}: Fjerner strekmarkeringer på x-aksen.
\item \code{fjern_y_ticks()}: Fjerner strekmarkeringer på y-aksen.
}

Fjerning av panellinjer kan være nyttig for å redusere støy i figurer
hvor det ikke er like stort behov for følgelinjer på begge aksene.
Fjerning av strekmarkeringer kan være nyttig for søylediagram med
kategoriske verdier, der strekmarkeringen er unødvendig.
Se eksempel under før hvordan figurene blir seende ut med de ulike
funksjonene.
}
\examples{
# Pakke for å laga figurar
library(ggplot2)

p = ggplot(iris, aes(x = Petal.Width, y = Petal.Length)) +
  geom_point()

# Eksempel på hva de ulike funksjonene gjør
p
p + fjern_x()
p + fjern_y()
p + fjern_x_ticks()
p + fjern_y_ticks()

# Flere av funksjonene kan brukes samtidig
p + fjern_x() +
  fjern_y_ticks()
}
