% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figur-funksjoner.R
\name{flytt_opp}
\alias{flytt_opp}
\title{Flytt-opp funksjon.}
\usage{
flytt_opp(y, tekst, hoyde = 0.015)
}
\arguments{
\item{y}{y-koordinat til (midten av) tekstane}

\item{tekst}{teksten i tekstane (berre brukt til å telja kor mange linjer det er).}

\item{hoyde}{høgda kvar linje tekst tar opp (i grafkoordinatar)}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Funksjon som flytter opp labels inni grafer hvis de kolliderer
}
