% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fodselsnummer-funksjoner.R
\name{fnr_er_gyldig}
\alias{fnr_er_gyldig}
\title{Sjekk om fødselsnummmer eller liknande er gyldige}
\usage{
fnr_er_gyldig(nummer, gyldige_typar = c("FNR", "H", "D", "FH"))
}
\arguments{
\item{nummer}{Tekstvektor med nummer som skal sjekkast.}

\item{gyldige_typar}{Tekstvektor med kva nummertypar verdiane
i \code{nummer} skal sjekkast mot. Må vera éin eller fleire av
\code{c("FNR", "H", "D", "FH")}.
Som standard vert nummera sjekka mot alle desse.}
}
\value{
Logisk vektor med \code{TRUE} eller \code{FALSE} for kvart element
i \code{nummer}, alt ettersom det høvesvis er eit gyldige nummer
av typane i \code{gyldige_typar} eller ikkje.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Tek inn vektor med \code{nummer} som skal sjekkast og \code{gyldige_typar},
som definerer kva type nummer desse potensielt kan vera
(fødselsnummer, H-nummer, D-nummer og liknande).
Gjev ut logisk vektor med \code{TRUE} for gyldige nummer og
\code{FALSE} for ugyldige nummer.
}
\details{
Dei moglege verdiane i \code{gyldige_typar} tyder:
\itemize{
\item \code{"FNR"}: Fødselsnummer.
Vanleg ellevesifra fødselsnummer
som inneheld fødselsdato (seks siffer) etterfølgd av personnummer
(fem siffer, der dei to siste er sjekksiffer).
\item \code{"D"}: D-nummer.
Ellevesifra nummer,
som fødselsnummer,
men modifisert ved at det er lagt til 4 på det fyrste sifferet.
\item \code{"H"}: H-nummer (hjelpenummer).
Ellevesifra nummer,
som fødselsnummer,
men modifisert ved at det er lagt til 4 på det tredje sifferet.
\item \code{"FH"}: FH-nummer (felles hjelpenummer).
Ellevesifra nummer der fyrste siffer er 8 eller 9,
dei neste åtte siffera er tilfeldige,
og dei to siste er kontrollsiffer.
}

Du kan lesa meir om dei ulike nummertypane på desse nettsidene:
\itemize{
\item \href{https://no.wikipedia.org/wiki/F\%C3\%B8dselsnummer}{Wikipedia: Fødselsnummer}
\item \href{https://lovas.info/2013/12/01/identitetsnummer-i-norge/}{Identitetsnummer i Norge}
}
}
\examples{
nummer = c(
  "15076500565", # Gyldig F-nummer
  "70019950032", # Gyldig D-nummer
  "01410199935", # Gyldig H-nummer
  "88888888831", # Gyldig FH-nummer
  "98019800546" # Ugyldig nummer generelt
)

fnr_er_gyldig(nummer)

fnr_er_gyldig(nummer, gyldige_typar = c("FNR", "H"))
}
\seealso{
Viss du vil ha ut type ID-nummer, kan du bruka \code{\link[=finn_type_idnummer]{finn_type_idnummer()}}.
}
