% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fodselsnummer-funksjoner.R
\name{fnr_foresla}
\alias{fnr_foresla}
\title{Foreslå lignende fødselsnummer}
\usage{
fnr_foresla(x)
}
\arguments{
\item{x}{Tekststreng med ugyldige fødselsnummer som det skal
foreslås lignende fødselsnummer for.}
}
\value{
Returnerer en tekstvektor med et gyldig fødselsnummer hvis det finnes et
nærliggende fødselsnummer som er gyldig. Hvis det ikke finnes et gyldig
fødselsnummer returneres en tom tekststreng.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Hvis det finnes ugyldige fødselsnummer kan denne funksjonen brukes for å
foreslå lignende fødselsnummer som \emph{er} gyldige. Dette gjøres ved å bytte
ut ett vilkårlig siffer, eller bytter om på to nabosiffer.
}
\details{
OBS! \emph{Ikke} bruk resultatene av funksjonen direkte. Sjekk de oppgitte
fødselsnummer i DIPS eller andre kilder for å kontrollere at de viser til
rett person.
}
\examples{
feiltastet_fodselsnummer = "15076600565"
fnr_foresla(feiltastet_fodselsnummer)
}
