% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapwhale-styler-funksjonar.R
\name{if_for_while_part_requires_braces_rapwhale}
\alias{if_for_while_part_requires_braces_rapwhale}
\title{Check if if, for or while loop expression require braces}
\usage{
if_for_while_part_requires_braces_rapwhale(pd, key_token)
}
\arguments{
\item{pd}{A parse table.}

\item{key_token}{The token that comes right before the token that contains the expression
to be wrapped (ignoring comments).
For if and while loops, this is the closing "')'",
for a for-loop it's "forcond".}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This is the case if they are not yet wrapped into curly braces.
}
\details{
Lik \link[styler:if_for_while_part_requires_braces]{styler::if_for_while_part_requires_braces},
men der alle uttrykk krev krøllparentesar (ikkje berre multi-line).
}
\keyword{internal}
