% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logiske-tester.R
\name{impl}
\alias{impl}
\alias{\%impl\%}
\title{Sjekk om sann \code{a} impliserer sann \code{b}}
\usage{
impl(a, b)

a \%impl\% b
}
\arguments{
\item{a}{Logisk vektor.}

\item{b}{Logisk vektor som elementvis skal sjekkes mot \code{a}.}
}
\value{
Logisk vektor som elementvis sier om \code{a = TRUE} impliserer \code{b = TRUE}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Tar inn to logiske vektorer \code{a} og \code{b}, og gir ut en logisk vektor som
elementvis sier om \code{a = TRUE} impliserer \code{b = TRUE}.
}
\details{
Funksjonen tar inn to logiske vektorer \code{a} og \code{b}, sjekker elementvis om
\code{a = TRUE} impliserer \code{b = TRUE}.

Den gir ut en logisk vektor som er \code{TRUE} for hvert elementpar fra
\code{a} og \code{b} der \code{a = FALSE}, \code{a = NA}, eller både
\code{a} og \code{b = TRUE}, og \code{FALSE} ellers. Den gir altså ut \code{FALSE} for
elementpar som motstrider at \code{a = TRUE} impliserer \code{b = TRUE}.

Funksjonen tilsvarer altså \code{!a | b}, men håndterer \code{NA} som \code{FALSE}.

Hvis \code{a} og \code{b} har ulik lengde blir den korteste brukt på ny nødvendig
antall ganger, likt som for logiske operatorer. Se \link[base:Logic]{base::Logic} for
detaljer.
}
\section{Functions}{
\itemize{
\item \code{a \%impl\% b}: \code{\link[=impl]{impl()}} som infiks-operator

}}
\examples{
library(tibble)

d = tribble(
  ~pas_id, ~bosted_by, ~bosted_bydel,
  1, "Bergen", "Åsane",
  2, "Bergen", "Landås",
  3, "Oslo", "Tjuvholmen",
  4, "Bergen", "Fana",
  5, "Oslo", "Haugenstua"
)

bosted_aasane = d$bosted_bydel == "Åsane"
bosted_bergen = d$bosted_by == "Bergen"

# Sjekk om bosted Åsane impliserer bosted Bergen
impl(bosted_aasane, bosted_bergen)

# Eller som infiks-operator
bosted_aasane \%impl\% bosted_bergen
}
