% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kodebok-valider.R
\name{kb_er_gyldig}
\alias{kb_er_gyldig}
\title{Tester om kodebok er gyldig}
\usage{
kb_er_gyldig(kb_glissen, sjekk_varnamn = TRUE, ...)
}
\arguments{
\item{kb_glissen}{Kodebok på glissen form, det vil si en kodebok som er
formatert i henhold til fagsenterets standardformat.}

\item{sjekk_varnamn}{Skal variabelnavn også sjekkes for gyldighet?
Godkjente variabelnavn er i snake-case; små bokstaver og "_" som mellomrom.}

\item{...}{Eventuelle argument \code{skjemaprefiks} og \code{fald_bokstavar},
brukt ved sjekking av variabelnamn.
Viss \code{skjemaprefiks} er sann, må òg namna vera
gyldige dersom me fjernar teksten fram til
og med første «_»-teikn.
Viss \code{fald_bokstavar} er sann, vert namna gjort
om til små bokstavar før testing, dvs. store bokstavar
og ei blanding av små og store bokstavar vert godtatt.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Funksjonen går gjennom en rekke tester for å se om en kodebok er gyldig. Tar inn en kodebok på glissen form.
}
