% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kodebok-fyll.R
\name{kb_fyll}
\alias{kb_fyll}
\title{Hent variabeltekst for kategoriske variabler}
\usage{
kb_fyll(df, kb, ..., .suffiks = "_tekst")
}
\arguments{
\item{df}{Datasett som inkluderer de kategoriske variablene.}

\item{kb}{Kodebok for registeret.}

\item{...}{Ytterligere argumenter.}

\item{.suffiks}{Suffiks for variabelnavn som kobles på det opprinnelige variabelnavnet.
Default er "_tekst".}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Funksjon for å hente variabeltekstar på kategoriske variablar
basert på kodebok.
Genererer nye variabler med oppgitt suffiks
og returnerer datasettet med de nye variablene inkludert.
For en variabel som er kodet med en vilkårlig verdi vil denne funksjonen
hente direkte fra kodebok hvilken gruppe de forskjellige
verdiene representerer.
}
\examples{
# Pakke for bruk av tibble-objekt og rør-operatoren
library(dplyr)

# eksempel kodebok
kb = tribble(
  ~variabel_id, ~verdi, ~verditekst,
  "kjonn", 1, "mann",
  "kjonn", 2, "kvinne",
  "med", 1, "antibac",
  "med", 2, "insulin",
  "med", 3, "ibux",
  "med", 4, "globoid",
  "gensp", 1, "Misfornøyd",
  "gensp", 2, "Både og",
  "gensp", 3, "Fornøyd"
)

# Eksempeldatasett
d = tribble(
  ~pasid, ~kjonn, ~alder, ~med, ~prem,
  101, 2, 18, 3, 2,
  102, 1, 37, 4, 2,
  103, 1, 17, 1, 3
)

# Fyll in verditekst for alle variabler i kodebok
kb_fyll(d, kb)

# Hvis det kun skal hentes verditekst for en gitt variabel
kb_fyll(d, kb, kjonn)

# Variabler kan kobles mot variabel_id med annet navn i kodebok
kb_fyll(d, kb, kjonn, prem = "gensp")

# Suffiks kan endres fra default "_tekst"
kb_fyll(d, kb, med, .suffiks = "_hei")
}
