% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_kb_v2.R
\name{kb_oqr_base_til_std}
\alias{kb_oqr_base_til_std}
\title{Konverter OQR kodebok fra basisformat til standard format.}
\usage{
kb_oqr_base_til_std(kb_oqr)
}
\arguments{
\item{kb_oqr}{Tar inn en kodebok på basisformat slik det leses inn fra les_kb_oqr_base}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Ved bruk av les_kb_oqr_base får vi inn en kodebok på et format som må endres
litt for å kunne brukes til å lese inn datadump.
Denne funksjonen gjør følgende endringer:
Kolonnenavn endres til våre standardnavn
Duplikate variabler reduseres innen hvert skjema
Statusvariabler utvides til å inneholde alle nivå
Obligatoriske variabler sjekkes for å se om de er aktiveringsspørsmål
Navn for variabeltyper konverteres til standardnavn
Standardkolonner velges ut
Unike skjemanavn tildeles basert på skjemaid
Se de underliggende funksjonene for mer detaljerte beskrivelser.
}
