% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kodebok-valider.R
\name{kb_til_kanonisk_form}
\alias{kb_til_kanonisk_form}
\title{Valider kodebok}
\usage{
kb_til_kanonisk_form(kb)
}
\arguments{
\item{kb}{Kodebok som skal valideres.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Gjer kodeboka om til kanonisk form, dvs. funksjonen fyller inn implisitte
verdiar for variablane i kodeboka.
}
