% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kjed_ord.R
\name{kjed_ord}
\alias{kjed_ord}
\title{Kjed saman elementa i ein vektor til ein tekststreng}
\usage{
kjed_ord(ord, skiljeteikn = ", ", og = " og ")
}
\arguments{
\item{ord}{Ein vektor med dei elementa som skal kjedast saman.}

\item{skiljeteikn}{Skiljeteikn som skal brukast mellom alle
elementa (utanom dei to siste).}

\item{og}{Tekst som skal setjast inn mellom dei to siste elementa.}
}
\value{
Ein tekststreng med eitt element,
der alle elementa frå \code{ord} er kjeda saman.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Tek inn ein vektor og kjedar saman (listar opp) elementa.
Returnerer resultatet som ein tekststreng.
}
\details{
Funksjonen er meint å brukast til å laga ei opplisting av
elementa i \code{ord}
(typisk ein tekst- eller talvektor)
på ein naturleg måte.

Funksjonen tek inn ein vektor og kjedar saman elementa med \code{skiljeteikn}
(som standard eit komma)
mellom kvart element, utanom dei to siste, som vert kjeda saman med \code{og}
(som standard ordet «og» med mellomrom rundt).
Resultatet vert returnert som ein tekststreng med eitt element.

Funksjonen handterar ulike typar inndata, som tekst, tal, dato
og andre typar objekt der \code{\link[base:as.character]{base::as.character()}} er definert.
Manglande verdiar vert gjort om til teksten \code{"NA"}.
}
\examples{
kjed_ord(c("Per", "Kari"))

kjed_ord(c("Per", "Kari", "Ola"))

kjed_ord(c("Per", "Kari", "Ola"), og = " & ")

kjed_ord(c("Førde", "Bergen og Voss", "Oslo", "Stavanger"),
  skiljeteikn = "; ",
  og = " – og til slutt "
)
}
