% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kompileringsfunksjonar.R
\name{kompiler_rnw}
\alias{kompiler_rnw}
\title{Kompiler .Rnw til .tex}
\usage{
kompiler_rnw(adresse)
}
\arguments{
\item{adresse}{Tekststreng med adresse til .Rnw-fila som skal kompilerast.}
}
\value{
Adressa til .tex-fila, viss den vart kompilert utan feil, usynleg.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Funksjonen kompilerer ei .Rnw-fil til .tex-fil
}
\details{
Dersom kompileringa ikkje vert fullført utan feil,
slettar funksjonen .tex-fila som svarar til \code{adresse}.
}
\examples{
\dontrun{
kompiler_rnw("H:/adresse/til/fil.Rnw")
}
}
