% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kompileringsfunksjonar.R
\name{kompiler_tex}
\alias{kompiler_tex}
\title{Kompiler .tex til .pdf}
\usage{
kompiler_tex(adresse, maksiter = 5, vis_feilmeldingar = TRUE)
}
\arguments{
\item{adresse}{Tekststreng med adresse til .tex-fila som skal kompilerast.}

\item{maksiter}{Heiltal for maks antal LuaLaTeX-køyringar. Standard = 5.}

\item{vis_feilmeldingar}{Logisk variabel som seier om feilmeldingar skal visast. Standard = TRUE.}
}
\value{
Adressa til .pdf-fila, usynleg.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Funksjonen kompilerer ei .tex-fil til .pdf-fil,
med maks \code{maksiter} LuaLaTeX-køyringar.
}
\examples{
\dontrun{
kompiler_tex("H:/adresse/til/fil.tex")
}
}
