% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_dd_oqr_csv.R
\name{konverter_boolske}
\alias{konverter_boolske}
\title{Konverter en variabel til logisk variabel}
\usage{
konverter_boolske(x, boolsk_usann, boolsk_sann, na_verdier = NA)
}
\arguments{
\item{x}{Vektor med verdier som skal konverteres til en logisk vektor.}

\item{boolsk_usann}{Hvilke verdier skal konverteres til FALSE? Kan inneholde flere verdier.}

\item{boolsk_sann}{Hvilke verdier skal konverteres til TRUE? Kan inneholde flere verdier.}

\item{na_verdier}{Hvilke verdier skal betraktes som NA? Kan inneholde flere verdier.}
}
\value{
En vektor som kun inneholder TRUE, FALSE eller NA.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Konverterer en vektor med verdier til en vektor som kun inneholder
TRUE, FALSE eller NA
}
\details{
Tar inn en vektor som skal konverteres, i tillegg til vektorer som
beskriver hva som skal regnes som TRUE, FALSE og NA.
}
\examples{
tallvektor = c(-1, 0, 1)
konverter_boolske(tallvektor,
  boolsk_usann = 0, boolsk_sann = 1,
  na_verdier = -1
)

tekstvektor = c("Ukjent", "Nei", "Ja")
konverter_boolske(tekstvektor,
  boolsk_usann = "Nei",
  boolsk_sann = "Ja", na_verdier = "Ukjent"
)

tekstvektor_flere_alt = c(NA, "Nei", "Ja", "-1", "0", "false")
konverter_boolske(tekstvektor_flere_alt,
  boolsk_usann = c("Nei", "0", "false"),
  boolsk_sann = c("Ja"), na_verdier = c(NA, "-1")
)
}
