% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_kb_v2.R
\name{konverter_oqr_kb}
\alias{konverter_oqr_kb}
\title{Hjelpefunksjon for konvertering av kodebok fra OQR-format til standardformat}
\usage{
konverter_oqr_kb(kb_mellom)
}
\arguments{
\item{kb_mellom}{kodebok på mellomformat}
}
\value{
kodebok på standardformat
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Tar inn kodebok på mellomformat og konverterer til standardformat.
Endrer navn på variabeltyper til standard navn
Velger standardkolonner som alle kodebøker skal ha på standardformat.
Tildeler unike skjemanavn for tabellene basert på skjemaid.
}
