% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_kb_v2.R
\name{konverter_tekst}
\alias{konverter_tekst}
\title{Konverter tekstvariabel til desimal eller dato}
\usage{
konverter_tekst(d, regex, parse_funksjon, ...)
}
\arguments{
\item{d}{tekst-vektor som skal konverteres}

\item{regex}{regex uttrykk for hvilke format det forventes at teksten innehar.}

\item{parse_funksjon}{en parse_*-funksjon fra readr pakken. Foreløpig kun støtte for parse_double og parse_date.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Funksjonen tar inn en tekst-vektor og konverterer til ønsket format.
Må oppgi et regex-uttrykk for å vise hvordan tekst-strengene som skal endres ser ut.
Funksjonen ser deretter etter mønsteret i tekst-vektoren og konverterer de det gjelder
til ønsket format basert på hvilken parse-funksjon som er oppgitt.
Verdier som ikke matcher regex vil endres til NA
}
\keyword{internal}
