% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kopier_latex_klassefil.R
\name{kopier_latex_klassefil}
\alias{kopier_latex_klassefil}
\title{Kopier LaTeX-klassefil til TeX Live eller liknande}
\usage{
kopier_latex_klassefil(texmappe = NULL)
}
\arguments{
\item{texmappe}{Adressa til mappa som LaTeX-klassefila skal kopierast
til. Viss \code{NULL} (standard), vert fila kopiert til
ei høveleg undermappe i TeX Live-mappa.}
}
\value{
Adressa til mappa som LaTeX-klassefila vart kopiert til, usynleg.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Kopierer LaTeX-klassefila for kvalreg-rapportar frå rapwhale til
TeX Live-mappa eller anna gjeven mappe.
}
\details{
Dersom \code{texmappe} er \code{NULL},
vert klassefila kopiert til undermappa
\verb{/tex/latex/kvalreg/} i TeX Live-mappa
definert av TeX-variabelen \code{TEXMFHOME}.
Det vert typisk noko slikt:
\verb{C:/Users/brukarnamn/texmf/tex/latex/kvalreg/}.
Dette gjer at ho automatisk vert funnen av LuaLaTeX.
Elles vert ho kopiert til den oppgjevne mappa \code{texmappe}.

Dersom mappa ikkje finst finst frå før, vert ho oppretta.
Dersom fila \code{kvalreg-rapport.cls} allereie finst i mappa det skal kopierast
til, vert ho skriven over.
}
\examples{
\dontrun{
kopier_latex_klassefil()

texmappe = tempdir()
kopier_latex_klassefil(texmappe)
}
}
