% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymiser.R
\name{lag_anonymiseringsfunksjon}
\alias{lag_anonymiseringsfunksjon}
\title{Lag funksjon for anonymisering av datavektorar}
\usage{
lag_anonymiseringsfunksjon(x, startnr = 1001)
}
\arguments{
\item{x}{Vektor (av valfri type/klasse) med verdiar som skal anonymiserast.}

\item{startnr}{Minste verdi for dei tilfeldige tala som skal genererast.}
}
\value{
Funksjon som tar inn ein vektor med verdiar
(som alle må inngå i \code{x})
og gjev ut ein heiltalsvektor med tilsvarande anonymiserte verdiar.
Han gjev ut feilmelding viss ein prøver
å bruka han på «ukjende» verdiar.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Lagar ein funksjon som kan brukast til å anonymisera verdiane i ein vektor
ved å byta dei ut med tilfeldige tal.
Funksjonen ein får ut,
er deterministisk, dvs. han gjev alltid ut same utverdi
(«tilfeldige» tal)
for same innverdi.
}
\details{
Denne funksjonen er meint å brukast når ein har verdiar
(for eksempel pasient-ID-ar)
som inngår i fleire datasett og som skal anonymiserast.
Då er det viktig at dei vert anonymiserte til same verdiar
på tvers av datasetta.

Sjå \code{\link[=anonymiser]{anonymiser()}} for informasjon om argumenta og detaljar
om anonymiseringa (desse er like i denne funksjonen).
}
\note{
Funksjonen må køyrast på ein vektor som inneheld \emph{alle}
verdiane ein i framtida ønskjer å anonymisera.
(Viss ein vil anonymisera pasient-ID-ar i eit kvalitetsregister,
er dette typisk data frå eit inklusjonsskjema eller eit pasientskjema).
Viss det ikkje finst ein slik vektor frå før,
må ein laga ein,
ved å samla alle verdiane ein vil anonymisera på tvers av datasett.
}
\examples{
library(tibble)
library(dplyr)

d_operasjon = tibble(
  pasient_id = c(2, 5, 7, 5, 8),
  sjukehus = c(
    "Bergen", "Oslo", "Stavanger",
    "Oslo", "Stavanger"
  )
)
d_oppfolging = tibble(
  pasient_id = c(7, 8),
  komplikasjonar = c(TRUE, FALSE)
)

anonymiser_pasids = lag_anonymiseringsfunksjon(d_operasjon$pasient_id)

d_operasjon_ano = mutate(d_operasjon, pasient_id = anonymiser_pasids(pasient_id))
d_oppfolging_ano = mutate(d_oppfolging, pasient_id = anonymiser_pasids(pasient_id))
d_operasjon_ano
d_oppfolging_ano
}
\seealso{
Bruk heller \code{\link[=anonymiser]{anonymiser()}} dersom du berre vil anonymisera éin vektor.
}
