% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figur-funksjoner.R
\name{lag_fig_histogram}
\alias{lag_fig_histogram}
\title{Lag histogram}
\usage{
lag_fig_histogram(d, x, binwidth = 1, boundary = 0, breaks_width = NULL, ...)
}
\arguments{
\item{d}{Datasett som inkluderer variabelen som skal brukast i histogrammet.}

\item{x}{Variabelen som skal plottast.}

\item{binwidth}{Søylebreidde (som i \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}). Standard er 1.}

\item{boundary}{Justeringsgrense for venstre søylekant (som i \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}).
Bør vanlegvis vera 0, som òg er standardverdi.}

\item{breaks_width}{Avstand/mellomrom som skal brukast mellom kvart aksetal.
Viss \code{NULL}, vert ein fornuftig standardverdi brukt.}

\item{...}{Ekstra argument som vert gjeve vidare til \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}.}
}
\value{
Eit histogram som ggplot-objekt.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Funksjon for å laga histogram.
}
\details{
Lagar eit vanleg histogram,
med \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} og \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}},
men med meir nyttige standardverdiar og enkelte visuelle forbetringar:
\itemize{
\item \code{boundary}-argumentet er som standard sett til 0,
slik at søyler med heiltalig \code{binwidth} alltid vil starta på eit heiltal.
\item Søylene starter heilt nede ved \emph{x}-aksen, og har 5 \% luft over seg.
\item Ein kan enkelt velja avstanden som skal vera mellom aksetala
med \code{breaks_width}.
\item Vassrette akselinjer vert ikkje vist.
}
}
\examples{
lag_fig_histogram(iris, Petal.Length, binwidth = .25, breaks_width = .5)
}
