% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figur-funksjoner.R
\name{lag_fig_linje}
\alias{lag_fig_linje}
\title{Lag linjegraf med 95 prosent konfidensintervall}
\usage{
lag_fig_linje(
  refline = NULL,
  refline_df = NULL,
  xlab = "År",
  ylab = NULL,
  angle = TRUE,
  konfint = TRUE,
  point_size = 2
)
}
\arguments{
\item{refline}{y-koordinaten til vassrett referanselinje}

\item{refline_df}{ev. dataramme med éi rad for kvart panel refline skal gjelda for}

\item{xlab}{tekst på x-aksen (standardverdi: "År")}

\item{ylab}{tekst på y-aksen (standardverdi: NULL (tom))}

\item{angle}{viss sann, vis verdiane på x-aksen på skrå (for å få plass til fleire)}

\item{konfint}{Legg til konfidensintervall på kvar punkt}

\item{point_size}{Storleik på punkt i grafen. Standardverdi er 2.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

description
}
\examples{
library(ggplot2)
d = data.frame("År" = c(2016, 2017, 2018), "Andel" = c(0.25, 0.5, 0.6))
p = ggplot(d, aes(
  x = År,
  y = Andel,
  ymin = c(0.2, 0.4, 0.5),
  ymax = c(0.3, 0.6, 0.7)
))

p + lag_fig_linje(konfint = FALSE, ylab = "Andel")
p + lag_fig_linje(refline = 0.55, konfint = FALSE, ylab = "Andel")
p + lag_fig_linje(refline = 0.55, konfint = TRUE, ylab = "Andel")
}
