% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figur-funksjoner.R
\name{lag_fig_shewhart}
\alias{lag_fig_shewhart}
\title{Lag shewhart-diagram}
\usage{
lag_fig_shewhart(
  d,
  y,
  x,
  figtype,
  nevner = NULL,
  tittel = NULL,
  gruppe = NULL,
  periode = NULL,
  x_navn = NULL,
  y_navn = NULL,
  ...
)
}
\arguments{
\item{d}{Datasett som inkluderer dei variablane som skal brukast shewhart-diagrammet.}

\item{y}{Vektor med tal som skal plottast på y-aksen (teljar).}

\item{x}{Vektor med verdiar som skal plottast på x-aksen, ofte datoar.}

\item{figtype}{Tekstvektor med kva type diagram som skal lagast. Sjå \code{\link[qicharts2:qic]{qicharts2::qic()}}
for tilgjengelege typar.}

\item{nevner}{Vektor med undergruppestorleikar (nemnar).}

\item{tittel}{Tekstvektor med tittel på plottet.}

\item{gruppe}{Eventuell variabel det skal grupperast på.}

\item{periode}{Tidsperiode brukt for å aggregera y-verdiar.
Til dømes \code{"month"} eller \code{"2 months"}.}

\item{x_navn}{Tekststreng med namnet på x-aksen.}

\item{y_navn}{Tekststreng med namnet på y-aksen.}

\item{...}{Eventuelle andre argument som vert gjeve vidare til \code{\link[qicharts2:qic]{qicharts2::qic()}}.}
}
\value{
Eit shewhart-diagram av typen \code{figtype}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
d = data.frame(
  dato = sample(
    seq.Date(
      from = as.Date("2021-01-01"),
      to = as.Date("2021-12-31"),
      by = "days"
    ),
    size = 1000,
    replace = TRUE
  ),
  avdeling = sample(
    x = c("HUS", "OUS", "UNN", "STA"),
    size = 1000, replace = TRUE
  ),
  teller = sample(
    x = c(TRUE, FALSE),
    size = 1000, replace = TRUE, prob = c(0.7, 0.3)
  ),
  nevner = sample(
    x = c(TRUE, FALSE),
    size = 1000, replace = TRUE, prob = c(0.9, 0.1)
  ),
  verdi = sample(100:250, size = 1000, replace = TRUE)
)

# Setter teller til FALSE og verdi til NA i de radene som nevner er FALSE
# slik det ville vært for en ekte indikatorfunksjon.
d$teller[!d$nevner] = FALSE
d$verdi[!d$nevner] = NA

lag_fig_shewhart(
  d = d, y = teller, x = dato, figtype = "p",
  nevner = nevner, gruppe = avdeling, periode = "2 months",
  x_navn = "Tidspunkt", y_navn = "Andel"
)

lag_fig_shewhart(
  d = d, y = verdi, x = dato, figtype = "xbar",
  nevner = nevner, gruppe = avdeling, periode = "2 months",
  x_navn = "Tidspunkt", y_navn = "Verdi"
)
}
