% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figur-funksjoner.R
\name{lag_fig_soyle_grunnplott}
\alias{lag_fig_soyle_grunnplott}
\title{Lag søylediagram-grunnplott}
\usage{
lag_fig_soyle_grunnplott(d, x, y, flip = TRUE, ...)
}
\arguments{
\item{d}{Datasett som inkluderer dei variablane som skal brukast søylediagrammet.}

\item{x}{Variabel for x-aksen - Vanlegvis er dette ein kategorisk variabel.}

\item{y}{Variabel for y-aksen - Ein kontinuerleg variabel. Kan vera prosent.}

\item{flip}{Logisk variabel som seier om plottet skal flippast. Standard verdi = \code{TRUE}.}

\item{...}{Ekstra argument som vert gjeve vidare til \code{\link[ggplot2:geom_col]{ggplot2::geom_col()}}.}
}
\value{
Eit søylediagram som ggplot-objekt.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Hjelpefunksjon for å laga søylediagram i \code{\link[=lag_fig_soyle]{lag_fig_soyle()}} og
\code{\link[=lag_fig_soyle_prosent]{lag_fig_soyle_prosent()}}.
}
\examples{
library(tibble)

d = tibble(gruppe = c("a", "b", "c"), verdi = c(2.6, 2.1, 3.2))
rapwhale:::lag_fig_soyle_grunnplott(d, gruppe, verdi)
rapwhale:::lag_fig_soyle_grunnplott(d, gruppe, verdi, flip = FALSE)
}
\keyword{internal}
