% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_dd_oqr_csv.R
\name{lag_formatspek}
\alias{lag_formatspek}
\title{Lag formatspek}
\usage{
lag_formatspek(
  skilletegn,
  desimaltegn,
  dato,
  klokkeslett,
  dato_kl,
  tidssone,
  tegnkoding,
  boolsk_sann,
  boolsk_usann,
  na_verdier
)
}
\arguments{
\item{skilletegn}{Angir hvilket tegn som brukes for å skille mellom kolonner i datafil}

\item{desimaltegn}{Angir hvilket tegn som brukes som desimaltegn? Må være ',' eller '.'}

\item{dato}{Angir hvilket dato-format som er brukt for datokolonner i inndata}

\item{klokkeslett}{Angir hvilket format som er brukt for klokkeslett}

\item{dato_kl}{Angir hvilket format som er brukt for variabler med både dato og klokkeslett}

\item{tidssone}{Angir hvilken tidssone som brukes}

\item{tegnkoding}{Angir tegnkodingen som brukes i inndata}

\item{boolsk_sann}{Angir hvilke verdier som tolkes som TRUE for boolske variabler}

\item{boolsk_usann}{Angir hvilke verdier som tolkes som FALSE for boolske variabler}

\item{na_verdier}{Angir hvilke verdier som tolkes som NA}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{lag_formatspek()} er en hjelpefunksjon som skal sikre at alle argumenter
i formatspek angis i riktig format og med riktig navn. Formatspek spesifiserer
hvilket format inndata er på og brukes i diverse innlesningsfunksjoner i rapwhale.
}
\keyword{internal}
