% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadump-valider.R
\name{lag_regelsett}
\alias{lag_regelsett}
\title{Lag regelsett}
\usage{
lag_regelsett(kodebok, oblig = TRUE, rekkefolge = TRUE)
}
\arguments{
\item{kodebok}{kodeboken datadumpen skal sjekkes mot.}

\item{oblig}{Argument for om obligatoriske felt skal sjekkes.}

\item{rekkefolge}{Argument for å sjekke om datadump har samme rekkefølge på
variabelnavn som det som er oppgitt i kodeboken.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Funksjon som automatisk lager et regelsett basert på informasjon fra
en kodebok. Brukes for å validere en datadump.
}
\keyword{internal}
