% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymse-funksjoner.R
\name{lag_tab_latex}
\alias{lag_tab_latex}
\title{Lag LaTeX-tabell}
\usage{
lag_tab_latex(dataframe, label, caption, wide = FALSE, ...)
}
\arguments{
\item{dataframe}{Dataramme som inneholder data som skal vises i tabell.}

\item{label}{Label for tabellen som brukes for å lage referanser i teksten.}

\item{caption}{Forklaringstekst for tabellen.}

\item{wide}{Valgmulighet for om tabellen skal være breiere enn
standard tekstbredde.
Må være \code{TRUE} eller \code{FALSE},
standard er \code{FALSE}.}

\item{...}{Ytterligere argumenter som kan gis til \code{\link[Hmisc:latex]{Hmisc::latex()}}.}
}
\value{
Tekst som, i en LaTeX-kompilator, genererer en tabell.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Lager pene LaTeX-tabeller i et standardisert format for alle årsrapporter,
med mulighet for å gjøre den stor nok til å dekke hele siden (\code{wide = TRUE}).
}
\details{
Ytterligere argumenter som kan gis til \code{\link[Hmisc:latex]{Hmisc::latex()}} inkluderer
blant annet \code{colhead = c()}.
}
\examples{
# Pakke for bruk av tibble-objekt og rør-operatoren
library(dplyr)

cars_top_mpg = mtcars |>
  arrange(desc(mpg)) |>
  head()
cars_top_mpg_tab = lag_tab_latex(cars_top_mpg,
  label = "mpg_table",
  caption = "Cars with best mileage"
)
cat(cars_top_mpg_tab)
}
