% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_valideringsdatasett.R
\name{lag_valideringsdatasett}
\alias{lag_valideringsdatasett}
\title{Lag valideringsdatasett}
\usage{
lag_valideringsdatasett(d_reg, indvars)
}
\arguments{
\item{d_reg}{Datasett (dataramme/tibble).}

\item{indvars}{Vektor med indeksvariablar.}
}
\value{
Valideringsdatasett som er klart til å fyllast ut med og sjekkast
opp mot data frå ein gullstandard (typisk pasientjournal).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Lagar eit valideringsdatasett med data frå eit registerdatasett klart til å
fyllast ut med og sjekkast opp mot data frå ein gullstandard (typisk
pasientjournal).
}
\details{
Funksjonen tek inn eit datasett \code{d_reg} og ein vektor \code{indvars} og
returnerer eit valideringsdatasett.

Datasettet \code{d_reg} skal vera anten data.frame eller tibble.

Vektoren \code{indvars} inneheld namna på alle indeksvariablane. Desse må alle
finnast som kolonnar i \code{d_reg}, og kombinasjonen av dei må unikt identifisera
kvar rad. Kolonnane i \code{d_reg} som ikkje er indeksvariablar er datavariablar
som skal validerast.

Valideringsdatasettet som vert gjeve ut vil ha éi rad for kvar datavariabel
for kvar rad i det opphavlege datasettet. Kvar rad i valideringsdatasettet
vil då innehalda verdien frå éin datavariabel i éi rad frå det opphavlege
datasettet. Denne verdien vil vera lagra i kolonnen \code{vld_verdi_intern_x}, der
\code{x} er typen til den aktuelle variabelen. Valideringsdatasettet  vil vera
på formatet som er definert i \code{\link[=er_valideringsdatasett_gyldig]{er_valideringsdatasett_gyldig()}}, med tomme
kolonnar \code{vld_verdi_ekstern_x}, klare til å fyllast ut og sidan sjekkast mot
verdiane i \code{vld_verdi_intern_x}.
}
\examples{
library(tibble)

# Eksempel på datasett frå eit register:
d_reg = tribble(
  ~pasid, ~dato_inn, ~dato_ut, ~vekt, ~hogd, ~biverk, ~biverk_hovud, ~biverk_mage, ~biverk_fot,
  5, as.Date("2020-06-07"), as.Date("2020-06-15"), 78, 183, TRUE, FALSE, TRUE, TRUE,
  5, as.Date("2020-12-13"), as.Date("2020-12-13"), 50, 179, TRUE, FALSE, TRUE, TRUE,
  7, as.Date("2020-08-09"), as.Date("2020-08-13"), 711, 196, TRUE, TRUE, TRUE, TRUE,
  13, as.Date("2021-01-05"), NA, NA, 163, FALSE, NA, NA, NA,
  14, as.Date("2021-01-05"), as.Date("2021-01-09"), 101, 182, TRUE, TRUE, FALSE, FALSE
)
d_reg

# Indeksvariablar:
indvars = c("pasid", "dato_inn")

d_vld = lag_valideringsdatasett(d_reg, indvars)
d_vld
}
