% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regn-sumskaar-generelt.R
\name{legg_til_eller_erstatt_kolonner}
\alias{legg_til_eller_erstatt_kolonner}
\title{Legg til / erstatt variabler i ett datasett med variabler i et annet}
\usage{
legg_til_eller_erstatt_kolonner(d_orig, d_ekstrakol)
}
\arguments{
\item{d_orig}{Originalt datasett (dataramme/tibble).}

\item{d_ekstrakol}{Dataramme/tibble som inneholder én eller flere
kolonner og samme antall rader som \code{d_orig}.}
}
\value{
Originalt datasett med ekstra kolonner lagt til / erstattet.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Legger variabler fra ett datasett til et annet datasett (med
overskriving av variabler som finnes i begge).
}
\details{
Variabler som finnes i \code{d_orig} fra før blir erstattet (med advarsel),
og nye blir lagt til helt til høyre.
}
\keyword{internal}
