% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regn-sumskaar-generelt.R
\name{legg_til_na_i_skaaringstabell}
\alias{legg_til_na_i_skaaringstabell}
\title{Legg til NA-rader i skåringstabell}
\usage{
legg_til_na_i_skaaringstabell(skaaringstabell)
}
\arguments{
\item{skaaringstabell}{Vanleg skåringstabell. Sjå definisjon
i \code{\link[=skaar_datasett]{skaar_datasett()}}.}
}
\value{
Skåringstabellen med \code{NA}-rader lagde til.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Legg til rader med både \code{verdi} og \code{koeffisient} lik \code{NA}, for
lettare skåring av datasett som har manglande verdiar.
}
\details{
For kvar eksisterande kombinasjon av \code{delskala} og \code{variabel} vert
det lagt til ei rad med både \code{verdi} og \code{koeffisient} sett til \code{NA}.
Dette vert berre gjort dersom ikkje alt finst ein \code{verdi} lik \code{NA}
(uavhengig av kva tilhøyrande \code{koeffisient} er).

Dette er berre ein intern hjelpefunksjon for å forenkla koden
for skåring av datasett. Med slike \code{NA}-rader vert skåren som vert
utrekna med \code{\link[=skaar_datasett_uten_validering]{skaar_datasett_uten_validering()}} automatisk til \code{NA} dersom
minst éin av variablane som inngår, manglar verdi (altså er \code{NA}).
}
\keyword{internal}
