% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_kb_v2.R
\name{legg_til_variabler_kb}
\alias{legg_til_variabler_kb}
\title{Legg til ekstra variabler i kodebok}
\usage{
legg_til_variabler_kb(kb_std, ekstra_data)
}
\arguments{
\item{kb_std}{kodebok på standardformat}

\item{ekstra_data}{Dataramme med variabler som skal legges til skjema.
Må inneholde skjema_id, skjemanavn, variabel_id, variabeletikett,
variabeltype, unik, obligatorisk og desimaler.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Funksjon for å legge til variabler i kodebok som finnes i datadump, men ikke
er med i kodebok.
Kolonnene skjema_id, skjemanavn, variabel_id, variabeletikett, variabeltype,
unik, obligatorisk og desimaler er obligatorisk.
Det er mulig å legge inn flere verdier gitt at det er gyldige kolonner
som er i kodebok fra før.
}
\keyword{internal}
