% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_dd_oqr_csv.R
\name{les_csv_base}
\alias{les_csv_base}
\title{Les inn csv-fil}
\usage{
les_csv_base(adresse, spesifikasjon, formatspek)
}
\arguments{
\item{adresse}{Filplassering for datadump som skal leses inn.}

\item{spesifikasjon}{Tibble med tre kolonner.
Inneholder varnavn_kilde, varnavn_resultat og vartype.}

\item{formatspek}{Liste som inneholder informasjon om hvordan csv-fil er spesifisert.
Gjelder desimaltegn, datoformat, klokkeslettformat,
tidssone, boolsk_sann, boolsk_usann.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{les_csv_base()} leser inn en csv-fil fra en filplassering.
Spesifisering av format og datastruktur gjøres i argumentene formatspek og spesifikasjon.
Se rapwhale::formatspek og rapwhale::spesifikasjon for eksempeloppsett for
ulike innstillinger.
Variabelnavn kan endres ved å inkludere en vektor med nye variabelnavn.
Aksepterte variabeltyper er: tekst, desimaltall, heltall, boolsk, dato, dato_kl og klokkeslett.
}
\keyword{internal}
