% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_csv_oqr.R
\name{les_csv_oqr}
\alias{les_csv_oqr}
\title{Les inn csv-fil med OQR formatspesifikasjon}
\usage{
les_csv_oqr(adresse, spesifikasjon)
}
\arguments{
\item{adresse}{en tekststreng som angir filplassering for csv-fil som
skal leses inn}

\item{spesifikasjon}{Tibble med tre kolonner. Inneholder varnavn_kilde,
varnavn_resultat og vartype.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{les_csv_oqr()} er en spesifikk versjon av \code{les_csv_base()} som er
tilpasset OpenQReg registerstruktur.
Funksjonen tar inn argumentene adresse og spesifikasjon. Adresse angir
filplassering, mens spesifikasjon angir opprinnelige og nye variabelnavn,
samt variabeltype.
Se rapwhale::spesifikasjon for eksempeloppsett for ulike innstillinger.
Variabelnavn kan endres ved å inkludere en vektor med nye variabelnavn.
Aksepterte variabeltyper er: tekst, desimaltall, heltall, boolsk, dato,
dato_kl og kl.
}
