% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dformat-checkware.R
\name{les_dd_checkware}
\alias{les_dd_checkware}
\title{Konverter Checkware-data til format fra kodebok}
\usage{
les_dd_checkware(
  mappe_dd,
  skjema_id,
  kontekst = c("T0", "T1", "T2"),
  dato = NULL,
  kodebok = NULL,
  valider_dd = TRUE,
  valider_kb = TRUE
)
}
\arguments{
\item{mappe_dd}{Adressa til datadump-mappa
(som inneheld éi undermappe, med namn på forma ÅÅÅÅ-MM-DD, for kvart uttak).}

\item{skjema_id}{ID til skjemaet ein vil henta inn
(brukt i filnamnet til datadumpen og i kolonnen «skjema_id» i kodeboka).}

\item{kontekst}{Berre ta med rader med kontekstverdi like disse (standard: T0, T1 eller T2).
Sett til NULL for å ta med alle.}

\item{dato}{Datoen ein skal henta ut kodeboka for (tekststreng eller dato).
Kan òg vera NULL, for å henta nyaste kodebok.}

\item{kodebok}{Kodebok på kanonisk form.
Kan òg vera NULL, og då vert kodeboka automatisk henta inn.}

\item{valider_dd}{Om man ønsker å validere datadumpen ja/nei (TRUE/FALSE).
Hvis man ønsker å ikke validere datadump med dd_er_gyldig,
kan man sette denne til FALSE. Default er TRUE.}

\item{valider_kb}{Om man ønsker å validere kodeboka ja/nei (TRUE/FALSE).
Hvis man ønsker å ikke validere kodebok med kb_er_gyldig,
kan man sette denne til FALSE. Default er TRUE.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Funksjon for å tilrettelegge checkware-data basert på kodebok,
hvor funksjonen automatisk henter inn kodebok som blir brukt (hvis ønskelig). \cr
Kodeboka vert brukt til å gje alle variablane rett format
(tal, tekst, dato, boolske/logiske verdiar osv.) og til å
sikra at datadumpen er i samsvar med kodeboka. \cr \cr
Som standard treng ein ikkje oppgje kodebok; ho vert automatisk henta inn.
Men dersom ein skal lesa inn mange skjema, er det lurare å lesa inn
kodeboka separat først, for at ting skal gå raskare (innlesing og validering
av kodeboka kan ta litt tid). Det er òg nødvendig å gjera det slik dersom
ein har kodeboka frå ei anna kjelde eller viss ein vil bruka ei modifisert
kodebok (generelt farleg!).

Returnerer et R-datasett for det aktuelle skjemaet,
med variabelnamn gjort om til ønnskede,
tilsvarende verdier funnet i kodeboka.
(I stedet for Q1, Q2, Q3 osv. som CheckWare ofte oppgir)
}
