% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dformat-mrs.R
\name{les_dd_mrs}
\alias{les_dd_mrs}
\title{Les datadump fra MRS-register}
\usage{
les_dd_mrs(mappe_dd, skjema_id, versjon = "Prod", dato = NULL, kodebok = NULL)
}
\arguments{
\item{mappe_dd}{Adressa til datadump-mappa
(som inneheld éi undermappe, med namn på forma ÅÅÅÅ-MM-DD, for kvart uttak).
Antagelse ligger til grunn at nyeste kodebok ligger i samme mappe
som de nyeste datadumpene.}

\item{skjema_id}{ID til skjemaet ein vil henta inn
(brukt i filnamnet og i kolonnen «tabell» i kodeboka).}

\item{versjon}{Om datadumpen er "Prod" eller om den er fra "QA". Standardverdi er "Prod".}

\item{dato}{Datoen ein skal henta ut kodeboka for (tekststreng eller dato).
Kan òg vera NULL, for å henta nyaste kodebok.}

\item{kodebok}{Kodebok på kanonisk form.
Kan òg vera NULL, og då vert kodeboka automatisk henta inn.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Les inn MRS-data frå gitt skjema ved hjelp av kodebok. \cr
Kodeboka vert brukt til å gje alle variablane rett format
(tal, tekst, dato, boolske/logiske verdiar osv.) og til å
sikra at datadumpen er i samsvar med kodeboka. \cr \cr
Som standard treng ein ikkje oppgje kodebok; ho vert automatisk henta inn. \cr
Men dersom ein skal lesa inn mange skjema, er det lurare å lesa inn
kodeboka separat først, for at ting skal gå raskare (innlesing og validering
av kodeboka kan ta litt tid). \cr
Det er òg nødvendig å gjera det slik dersom ein har kodeboka frå ei anna
kjelde eller viss ein vil bruka ei modifisert
kodebok (generelt farleg!).

Returnerer et R-datasett for det aktuelle skjemaet.
}
