% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dformat-checkware.R
\name{les_kb_checkware}
\alias{les_kb_checkware}
\title{Konverter Checkware kodebok til standardformat}
\usage{
les_kb_checkware(mappe_dd, dato = NULL, valider_kb = TRUE)
}
\arguments{
\item{mappe_dd}{Adressa til datadump-mappa
(som inneheld éi undermappe, med namn på forma ÅÅÅÅ-MM-DD, for kvart uttak).
Her er det gitt at nyeste kodebok legges i samme mappe
som nedhentede datadumper.}

\item{dato}{Hvis man ønsker å hente kodebok fra en spesifikk dato.
Hvis ikke hentes dette fra nyeste dato. Default er NULL.}

\item{valider_kb}{Om man ønsker å validere kodeboka ja/nei (TRUE/FALSE).
Hvis man ønsker å ikke validere kodebok med kb_er_gyldig,
kan man sette denne til FALSE. Default er TRUE.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Les inn OQR-kodebok på dokumentert format og gjer om til vårt standardformat (kanonisk form).

Returnerer kodeboken på standardformat (kanonisk form), med variabelnamn gjort om til små bokstavar.
}
