% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dformat-mrs.R
\name{les_kb_mrs}
\alias{les_kb_mrs}
\title{Konverter MRS-kodebok til standardformat}
\usage{
les_kb_mrs(mappe_dd, dato = NULL)
}
\arguments{
\item{mappe_dd}{Adressa til datadump-mappa
(som inneheld éi undermappe, med namn på forma ÅÅÅÅ-MM-DD, for kvart uttak).
Her er en antagelse at nyeste versjon av kodeboka ligger
i samme mappe som datadumpene.}

\item{dato}{Datoen ein skal henta ut kodeboka for (tekststreng eller dato).
Kan òg vera NULL, for å henta nyaste kodebok.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Gjer om MRS-kodebok til kodebok på standardformat.

Returnerer kodeboka på standardformat (kanonisk form), med variabelnamn gjort om til små bokstaver.
}
