% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dformat-oqr.R
\name{les_kb_oqr}
\alias{les_kb_oqr}
\title{Konverter OQR-kodebok til standardformat}
\usage{
les_kb_oqr(mappe_dd, reg_id, dato = NULL, valider_kb = TRUE)
}
\arguments{
\item{mappe_dd}{Adressa til datadump-mappa
(som inneheld éi undermappe, med namn på forma ÅÅÅÅ-MM-DD, for kvart uttak).}

\item{reg_id}{ID som identifiserer registeret og er prefiks til alle filnamna.}

\item{dato}{Datoen ein skal henta ut kodeboka for (tekststreng eller dato).
Kan òg vera NULL, for å henta nyaste kodebok.}

\item{valider_kb}{Skal kodeboka automatisk validerast?
Ho må då vera gyldig for at ein skal få noko ut.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Funksjonen leser inn OQR-kodebok på dokumentert format
og gjer om til vårt standardformat (kanonisk form).

Returnerer kodeboka på standardformat (kanonisk form),
med variabelnamn gjort om til små bokstavar.
}
