% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_kb_v2.R
\name{les_kb_oqr_base}
\alias{les_kb_oqr_base}
\title{Les inn OQR-kodebok}
\usage{
les_kb_oqr_base(adresse)
}
\arguments{
\item{adresse}{filplassering for kodebok.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

les_kb_oqr_base leser inn en base-versjon av OQR-kodebok. Det gjøres ingen
validering eller korreksjoner på kodeboken her. Den leses inn slik den er.
Eneste argument er adresse, som angir filplassering for kodebok.

Returnerer en OQR-kodebok som inneholder alt av informasjon, men som har
en del mangler. Variabeltyper er gjerne feil og tall- og  dato-kolonner
kan inneholde tekst-verdier.
}
\keyword{internal}
