% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_kb_v2.R
\name{les_kb_oqr_v2}
\alias{les_kb_oqr_v2}
\title{Funksjon for å lese inn kodebok for OQR-register}
\usage{
les_kb_oqr_v2(adresse, valider = TRUE)
}
\arguments{
\item{adresse}{en tekststreng for filplassering for kodebok.}

\item{valider}{TRUE eller FALSE for å velge om kodebok skal valideres ved innlesning.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Hovedfunksjon for å lese inn kodebok for OQR-register.
Tar inn filplassering for kodebok og argumentet valider som angir om det skal
gjøres validering av kodebok ved innlesning.
}
