% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_dd_oqr_csv.R
\name{les_varnavn}
\alias{les_varnavn}
\title{Les inn variabelnavn fra datafil}
\usage{
les_varnavn(adresse, formatspek)
}
\arguments{
\item{adresse}{Filplassering for datafil som skal leses inn. Henter kun ut variabelnavn.}

\item{formatspek}{Liste med formatspesifikasjon.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{les_varnavn()} henter ut første rad fra en datafil. Bruker filplassering som argument.
}
\details{
Returnerer en vektor med variabelnavn slik de er i datafilen.
}
