% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymse-funksjoner.R
\name{normaliser_varnamn}
\alias{normaliser_varnamn}
\title{Normaliser variabelnamn}
\usage{
normaliser_varnamn(x)
}
\arguments{
\item{x}{Tekststreng med variabelnavn som skal konverteres.}
}
\value{
Returnerer en tekststreng av samme lengde som \code{x},
hvor alle elementer er skrevet i "snake_case".
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Funksjon for å konvertere variabelnavn til standardformat.
}
\details{
Standardformatering er "snake_case",
hvor "_" blir brukt som skilletegn,
og kun små bokstaver benyttes.
Om variabelnavnene som blir tatt inn produserer flere like variabelnavn
skrevet i "snake_case" vil man motta en advarsel.
}
\note{
R tillater variabelnavn uten bokstaver,
for eksempel "._.".
Det gir ikke mening å skrive et slikt variabelnavn i "snake_case".
Derfor tillater \code{\link[=normaliser_varnamn]{normaliser_varnamn()}}
kun variabelnavn som inneholder minst en bokstav.
}
\examples{
normaliser_varnamn(
  c("hopp og.SprettTest", "SykdomsAktivitet.PasientGlobalSykdomsaktivitet")
)
}
