% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapporttekst-funksjoner.R
\name{num}
\alias{num}
\title{Konverter tal til LaTeX-format}
\usage{
num(x, desimalar = NULL)
}
\arguments{
\item{x}{Vektor med tala ein ønskjer å få ut på LaTeX-format.}

\item{desimalar}{Kor mange desimalar som skal visast etter kommaet.}
}
\value{
Vektor med LaTeX-kode.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Tar inn ein vektor med tal og gjev ut ein vektor
med LaTeX-kode for finformatert vising av tala,
for eksempel med tusenskiljeteikn
og med avrunding til eit fast tal desimalar.
}
\details{
Tala vert viste med tusenskiljeteikn («12 345», ikkje «12345»),
desimalteiknet vert
(viss språket er norsk)
komma («3,14»),
ikkje punktum («3.14»),
negative tal vert viste med ekte minusteikn («−42»),
ikkje bindestrek («-42»),
\code{NA}-verdiar vert viste som ein kort tankestrek («–»)
og \code{NULL}-verdiar returnerer ein tom tekstvektor.

Tala vert avrunda til \code{desimalar} desimalar
med den vanlege avrundingsregelen i R.
Viss \code{desimalar} er NULL,
vert det vist så mange desimalar som
\code{\link[=format]{format()}}-funksjonen viser som standard.
(Men,
i motsetning til den funksjonen,
vert tala aldri viste i eksponentiell notasjon.)

Merk at funksjonen berre skal brukast på vanlege tal,
ikkje årstal,
fødselsnummer eller liknande.

Det vert brukt ein funksjon gjort tilgjengeleg av
\code{kvalreg-rapport}-klassen
for formatering av tala.
Nøyaktig kva funksjon som vert brukt,
kan endrast i framtida,
men funksjonaliteten vil vera lik.
}
\note{
Viss \code{desimalar} er oppgjeve,
vert det alltid vist \emph{nøyaktig} så mange desimalar,
sjølv om dei siste vert 0.
}
\examples{
# Til bruk i setningar i LaTeX
paste0("Store tal som ", num(123456789), " får mellomrom som tusenskiljeteikn.")
paste0("Pi avrunda til fire desimalar er ", num(pi, desimalar = 4), ".")
paste0("Nulldesimalar vert òg viste: ", num(12, desimalar = 2))
}
