% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regn-sumskaar-generelt.R
\name{oppsummer_ugyldige_verdier}
\alias{oppsummer_ugyldige_verdier}
\title{Presenter ugyldige verdier i et datasett på en god måte}
\usage{
oppsummer_ugyldige_verdier(d_ugyldige)
}
\arguments{
\item{d_ugyldige}{Tibble på formatet gitt ut av
\code{\link[=finn_ugyldige_verdier]{finn_ugyldige_verdier()}}.}
}
\value{
Tekststreng som inneholder variabelnavn og tilhørende
feilverdier (sortert alfabetisk etter variabelnavn og så
rekkefølge i \code{d_ugyldige$radnr}). Hvis det ikke finnes ugyldige
verdier, returneres tekststrengen \code{"Alle verdiene er gyldige"}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Gir ut en oversiktlig fremstilling av tibble-en som returneres av
\code{\link[=finn_ugyldige_verdier]{finn_ugyldige_verdier()}}.
}
\keyword{internal}
