% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_kb_v2.R
\name{oqr_til_std_variabeltyper}
\alias{oqr_til_std_variabeltyper}
\title{Konverter navn for variabeltyper fra OQR til standard}
\usage{
oqr_til_std_variabeltyper(kb_mellom)
}
\arguments{
\item{kb_mellom}{Kodebok på mellomformat}
}
\value{
kodebok på mellomformat, men med standardnavn for variabeltyper
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Endrer navn for variabeltyper fra OQR-navn til standardnavn.
Sjekker også at det ikke er ukjente variabeltyper i kodebok.
}
\keyword{internal}
