% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/del-aar.R
\name{periode_til_tidslinje}
\alias{periode_til_tidslinje}
\title{Periode til tidslinje}
\usage{
periode_til_tidslinje(aar, delnummer, antall_deler)
}
\arguments{
\item{aar}{Vektor med årstall (på heltallsformat)}

\item{delnummer}{Heltallsvektor som angir hvilken periode datapunktet tilhører}

\item{antall_deler}{Heltall som angir hvor mange like deler året skal deles
inn i.}
}
\value{
Gir ut en vektor med desimaltall som representerer kordinatene for de ulike
punktene i inndata fordelt på en akse.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Tar inn en vektor med årstall, en vektor med periode-tilhørighet og et
heltall som sier hvor mange deler året skal deles inn i.
}
\details{
Funksjonen legger til en desimaltallsverdi til \code{aar},
som blir bestemt av \code{delnummer} og \code{antall_deler}.
Argumentet \code{antall_deler} forteller hvor mange intervall man ønsker
å dele året inn i.
Om \code{delnummer} er 1,
betyr det da at dette
\code{aar}-elementet skal hører til i det første intervallet.
Hvert intervall blir representert med midtpunktet i intervallet.
Dermed,
om man velge å dele året inn i kvartaler,
blir det første punktet lik \code{2019.125},
hvor \code{aar = 2019}.
}
\note{
Om man har datoer med måned og dag,
ikke kun årstall,
kan man eventuelt bruke \code{\link[=tid_til_tidslinje]{tid_til_tidslinje()}},
som finner \code{delnummer} automatisk.
}
\examples{
aarstall = c(2014, 2015, 2016, 2017)
delnummer = c(1, 2, 2, 4)
antall_deler = 4
periode_til_tidslinje(aarstall, delnummer, antall_deler)
}
