% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapporttekst-funksjoner.R
\name{prosent}
\alias{prosent}
\title{Vis prosent-verdi som LaTeX-tekst}
\usage{
prosent(x, desimalar = 0)
}
\arguments{
\item{x}{Vektor med tallene en ønsker å konvertere til prosent og få ut på LaTeX-format.}

\item{desimalar}{Antall desimaler som skal vises etter kommaet.}
}
\value{
Vektor med LaTeX-kode.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Tar inn en vektor med tall som konverteres til prosent
og gir ut en vektor med LaTeX-kode for finformatert
visning av tallene med prosenttegn.
Tallene blir avrundet til et bestemt antall desimaler.
}
\details{
Tallene blir konvertert til prosent og blir vist med LaTeX-kode
for formatering til prosenttegn
(«0.1234» blir «12.34 \%»).

Funksjonen \code{\link[=num]{num()}} blir brukt til finformatering av tallene.
Funksjonaliteten til \code{num} er gitt i \link[=num]{beskrivelsen av funksjonen}.
}
\examples{
menn = 5
kvinner = 7
andel_menn = menn / (menn + kvinner)

# Til bruk i setning i latex
paste0("Andel menn er ", prosent(andel_menn), ".")
paste0("Andel menn med 2 desimaler er ", prosent(andel_menn, desimalar = 2), ".")
}
