% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_kb_v2.R
\name{reduser_duplikate_variabler}
\alias{reduser_duplikate_variabler}
\title{Reduser duplikate variabler}
\usage{
reduser_duplikate_variabler(kb_mellom)
}
\arguments{
\item{kb_mellom}{kodebok på mellomformat}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Funksjon for å trekke ut unike variabler innenfor et skjema.
Foreløpig er denne funksjonen laget spesifikt for les_kb_v2.R, men den kan kanskje
utvides til å dekke mer generelle tilfeller.
Tar inn en kodebok på mellomformat og returnerer en kodebok på mellomformat
som har unike variabler innenfor hver tabell.
Grunnen til at vi gjør det er at enkelte OQR-register har flere skjema som registreres
i samme tabell, for eksempel oppfølging 1år, oppfølging 2år osv.
Vi vil kun ha en variabel som gjelder for alle disse skjema innen samme tabell.
}
\keyword{internal}
