% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymse-funksjoner.R
\name{regn_konfint_bin}
\alias{regn_konfint_bin}
\title{Konfidensintervall for binomisk fordeling}
\usage{
regn_konfint_bin(x, n, konf_niva = 0.95, alfa = lifecycle::deprecated())
}
\arguments{
\item{x}{Antall suksesser i forsøket.}

\item{n}{Antall uavhengige forsøk.}

\item{konf_niva}{Konfidensnivå.
Standardverdi er 0.95, som tilsvarer et 95 \%-konfidensintervall.}

\item{alfa}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Utdatert og erstatta av \code{konf_niva} (tilsvarande \code{1 - alfa}).
Vert fjerna i neste versjon av pakken.}
}
\value{
Returnerer en tibble med nedre og øvre grense for et
Wilson-konfidensintervall på angitt nivå.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Brukar Wilson-intervallet, som anbefalt i «Binomial confidence intervals
and contingency tests: mathematical fundamentals and the evaluation of
alternative methods», av Sean Wallis, University College London.
}
\examples{
n_forsok = 1000
n_suksess = sample.int(n_forsok, 1)
regn_konfint_bin(n_suksess, n_forsok)
}
