% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymse-funksjoner.R
\name{regn_konfint_univar}
\alias{regn_konfint_univar}
\title{Konfidenstinervall basert på gjennomsnittet til en kontinuerlig variabel}
\usage{
regn_konfint_univar(x, bootstrap = FALSE, konf_niva = 0.95, R = 9999)
}
\arguments{
\item{x}{Numerisk vektor man ønsker å lage et konfidensintervall for.}

\item{bootstrap}{Logisk verdi for om man ønsker å bruke bootstrap-metoden
eller ei.
Standardverdi er \code{FALSE},
altså å ikke bruke bootstrap.}

\item{konf_niva}{Numerisk verdi som angir konfidensnivået.
Standardverdi er 0.95,
som tilsvarer et 95 \%-konfidensintervall.}

\item{R}{Antall bootstrap-replikasjoner.
Kun i bruk når \code{bootstrap} er \code{TRUE}.
Standardverdi = 9999.}
}
\value{
Returnerer en tibble med nedre og øvre grense for konfidensintervall,
samt gjennomsnittsverdien av \code{x}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Finner konfidenstinervall basert på gjennomsnittet til en
kontinuerlig variabel med mulighet for bootstrap lagt inn i funksjonen.
}
\details{
Inndatasettet kan enten være en vektor,
eller en kolonne i en tibble/dataramme.
Om man anvender en tibble/dataramme,
kan man for eksempel bruke funksjonen inni \code{\link[dplyr:summarise]{dplyr::summarise()}}.
Hvis inndataene er gruppert,
blir resultatet regnet ut på gruppenivå,
med én rad per gruppe.

Funksjonen ignorerer \code{NA}-verdier i \code{x}.
}
\examples{
regn_konfint_univar(runif(100))

# Konfidensintervall for drivstofforbruk, gruppert på type girkasse
library(dplyr)

mtcars |>
  group_by(am) |>
  summarise(regn_konfint_univar(mpg))
}
