% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samanliknarar.R
\name{samanlikn_identisk}
\alias{samanlikn_identisk}
\title{Undersøk om verdiar i vektorar er identiske}
\usage{
samanlikn_identisk(verdi1, verdi2, varnamn = NULL)
}
\arguments{
\item{verdi1}{Vektor med verdiar av valfri type.}

\item{verdi2}{Vektor med verdiar av same type som \code{verdi1}.}

\item{varnamn}{Vektor med namn på variablane som skal samanliknast.
Må ha same lengd som \code{verdi1} og \code{verdi2} eller vera \code{NULL}.}
}
\value{
Logisk vektor som elementvis seier om \code{verdi1} er identisk
med \code{verdi2} (\code{TRUE}) eller ikkje (\code{FALSE}). Vil aldri innehalda
\code{NA}-verdiar.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Samanliknar elementvis om verdiane i to vektorar er identiske
(inkludert tilfellet der begge er \code{NA}). Er meint å brukast
saman med \code{\link[=analyser_valideringsdatasett]{analyser_valideringsdatasett()}}
}
\details{
Funksjonen samanliknar verdiane i \code{verdi1} og \code{verdi2} elementvis og
returnerer ein logisk vektor som seier om dei er identiske.
Fungerer i praksis som \code{==}, men med den forskjellen at to \code{NA}-verdiar
òg vert rekna som identiske.

Argumentet \code{varnamn} har ingen verknad, men er med for at funksjonen
skal vera på same format som andre samanliknarar, for kompatibilitet
med \code{\link[=analyser_valideringsdatasett]{analyser_valideringsdatasett()}}.
}
\examples{
# Samanlikn NA-handteringa med den i «==»
x1 = c(5, 7, NA, NA)
x2 = c(5, 8, NA, 5)
samanlikn_identisk(x1, x2)
x1 == x2

# Handterer verdiar av valfri type/klasse
samanlikn_identisk(
  verdi1 = as.Date(c("1999-05-17", "1999-12-24")),
  verdi2 = as.Date(c("1999-05-17", "2000-12-24")),
  varnamn = c("dato_ut", "dato_ut")
)
}
