% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_kb_v2.R
\name{sjekk_obligatorisk}
\alias{sjekk_obligatorisk}
\title{sjekker at obligatoriske variabler er synlige for bruker}
\usage{
sjekk_obligatorisk(kb_mellom)
}
\arguments{
\item{kb_mellom}{Kodebok på mellomformat}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Enkel sjekk for om obligatoriske variabler er aktiveringsspørsmål.
Hvis en obligatorisk variabel ikke er et aktiveringsspørsmål settes
obligatorisk til nei.
}
\keyword{internal}
