% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regn-sumskaar-generelt.R
\name{sjekk_skaaringstabell}
\alias{sjekk_skaaringstabell}
\title{Sjekk om skåringstabell er gyldig}
\usage{
sjekk_skaaringstabell(skaaringstabell)
}
\arguments{
\item{skaaringstabell}{Dataramme/tibble (skåringstabell).}
}
\value{
\code{NULL} (usynlig).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Gir feilmelding hvis skåringstabellen ikke er gyldig.
}
\details{
Sjekker at \code{skaaringstabell} inneholder riktige kolonner og riktige
variabeltyper. En gyldig skåringstabell skal ha fire kolonner,
\code{delskala} (tekst), \code{variabel} (tekst), \code{verdi} (numerisk) og
\code{koeffisient} (numerisk), og det skal bare finnes én rad
per kombinasjon av \code{delskala}, \code{variabel} og \code{verdi}.
Kolonnene \code{delskala} og \code{koeffisient} kan ikke ha \code{NA}-verdier.
Viss en rad har verdien \code{NA} i kolonnen \code{variabel},
\emph{må} den også ha verdien \code{NA} i kolonnen \code{verdi}.

I tillegg må hver variabel som inngår i en gitt delskala,
ha oppføringer for alle ikke-\code{NA}-verdiene som variabelen
kan ta i de \emph{andre} delskalaene den inngår i
(det vil si, har minst én ikke-\code{NA}-verdi i).

Gir ut feilmelding hvis \code{skaaringstabell} er ugyldig.
}
\keyword{internal}
