% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regn-sumskaar-generelt.R
\name{sjekk_variabelnavn}
\alias{sjekk_variabelnavn}
\title{Sjekk at oppgitte variabler finnes i datasettet}
\usage{
sjekk_variabelnavn(d, variabelnavn)
}
\arguments{
\item{d}{Dataramme/tibble.}

\item{variabelnavn}{Vektor med variabelnavn (som skal sjekkes om
finnes i \code{d}).}
}
\value{
\code{NULL} (usynlig).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Tar inn et datasett og en vektor med variabelnavn. Gir
feilmelding hvis datasettet ikke inneholder alle variabelnavnene i
vektoren.
}
\details{
Gir feilmelding hvis \code{d} ikke inneholder alle variabelnavnene
i \code{variabelnavn}. Feilmeldingen oppgir hvilke variabelnavn som
eventuelt mangler.
}
\keyword{internal}
