% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regn-sumskaar-generelt.R
\name{sjekk_variabelverdier}
\alias{sjekk_variabelverdier}
\title{Sjekk at verdier i et datasett er gyldige}
\usage{
sjekk_variabelverdier(d, verditabell, godta_manglende)
}
\arguments{
\item{d}{Dataramme/tibble som kun inneholder kolonner med identiske
navn som i \code{verditabell$variabel}.}

\item{verditabell}{Dataramme/tibble med to kolonner (\code{variabel}
og \code{verdi}) som sier hvilke verdier som er gyldige for hvilke
variabler. Variabel-kolonnen må være av typen tekst og
verdi-kolonnen må være numerisk.}

\item{godta_manglende}{Skal manglende verdier i
i \code{d} godtas (standard \code{FALSE})?
Hvis ikke, blir det gitt ut
en feilmelding dersom det finnes manglende verdier (\code{NA}-verdier)
som ikke har tilhørende oppføring i \code{verditabell}.}
}
\value{
\code{NULL} (usynlig).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Gir feilmelding dersom verdiene i datasettet ikke er i samsvar
med en verditabell som sier hvile verdier som er gyldige for
hvilke variabler.
}
\details{
Gir feilmelding hvis \code{d} inneholder en verdi som ikke har
tilhørende oppføring av \code{variabel} og \code{verdi} i \code{verditabell}.
Feilmeldingen inneholder oversikt over antall ugyldige verdier,
samt hvilke variabler og verdier dette gjelder.

Gir også feilmelding hvis \code{verditabell} ikke er \code{tibble}/\code{data.frame}
og/eller mangler minst én av kolonnene \code{variabel} og \code{verdi},
eller hvis \code{d} inneholder verdier som ikke er numeriske.
}
\keyword{internal}
