% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regn-sumskaar-generelt.R
\name{skaar_datasett}
\alias{skaar_datasett}
\title{Skår spørreskjema basert på skåringstabell}
\usage{
skaar_datasett(
  d,
  skaaringstabell,
  variabelnavn = NULL,
  godta_manglende = FALSE
)
}
\arguments{
\item{d}{Dataramme/tibble som inneholder spørreskjema-variabler +
eventuelt andre variabler. Spørreskjema-variablene må være
numeriske.}

\item{skaaringstabell}{Dataramme/tibble som sier hvordan
\code{d} skal skåres. Må ha fire kolonner,
\code{delskala} (tekst), \code{variabel} (tekst), \code{verdi} (numerisk) og
\code{koeffisient} (numerisk), og det kan bare finnes én rad
per kombinasjon av \code{delskala}, \code{variabel} og \code{verdi}.
Se detaljer nedenfor.}

\item{variabelnavn}{Navn på variabler i datasettet som ikke er
identiske med de standardiserte navnene i skåringstabellen. Bruk
syntaksen
\code{c(std_navn_1 = "dd_navn_1", std_navn_2 = "dd_navn_2")}.
Nye navn trenger kun oppgis for spørreskjema-variabler som har
avvikende navn fra skåringstabellen. Hvis \code{NULL}, blir
det antatt at alle navnene er i samsvar med skåringstabellen.}

\item{godta_manglende}{Skal manglende verdier (\code{NA}-verdier) i
spørreskjema-variablene i \code{d} godtas (som standard nei)?
Hvis ikke, blir det gitt ut
en feilmelding om det finnes manglende verdier.
Se detaljer nedenfor.}
}
\value{
Datasett likt \code{d}, men med sumskår(er) lagt til, eventuelt
erstattet. Nye sumskår-kolonner blir i utgangspunktet lagt til på
slutten av av \code{d}, i samme rekkefølge som i \code{skaaringstabell$delskala}.
Hvis \code{d} imidlertid alt innholder en variabel med samme navn som en
\code{delskala}, blir denne denne stående der den er, men overskrevet med
nyutregnet sumskår. Det blir i så fall gitt ut en advarsel.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Regner ut sumskår(er) i et datasett basert på en skåringstabell.
Sjekker også at alle verdiene i datasettet er i samsvar med
skåringstabellen.
}
\details{
Funksjonen regner ut vektede sumskårer ved hjelp av en skåringstabell.
Skåringstabellen sier hvordan de ulike svaralternativene skal inngå i
sumskårene. Et spørreskjema kan ha flere
sumskårer, som vi kaller \emph{delskalaer}. (For eksempel kan et skjema for
livskvalitet ha én delskala for fysisk livskvalitet og én for psykisk
livskvalitet.) For hver rad i \code{d} og hver \code{delskala} i \code{skaaringstabell}
blir sumskåren \code{delskala} i \code{d} lik summen av de tallene i \code{koeffisient}
der \code{verdi} for variabelen \code{variabel} i skåringstabellen er lik
verdien av tilhørende variabel i \code{d},
pluss et eventuelt konstantledd.

For å legge til et konstantledd for en sumskår legger man til en rad i
\code{skaaringstabell}.
I raden skal \code{delskala} være sumskåren konstantleddet skal gjelde for,
\code{variabel} og \code{verdi} begge være \code{NA},
og \code{koeffisient} være verdien til konstantleddet.

Merk at \code{verdi} også \emph{kan} være \code{NA} utenom konstantleddtilfellet.
Da blir manglende svar
(dvs. \code{NA-verdier}) for tilhørende variabel og delskala regnet som
gyldig svar og har en tilhørende \code{koeffisient} (som ikke er \code{NA}).
Dette kan være nyttig i tilfeller hvor svaralternativ som «Vet ikke»
og manglende svar begge blir kodet som \code{NA} og skal bidra til
sumskåren. Men merk at slik koding er en dårlig idé. Alle gyldige
svaralternativ bør ha eksplisitte verdier.

Hvis argumentet \code{godta_manglende} er satt til \code{FALSE}, vil en få
feilmelding dersom det finnes noen \code{NA}-verdier i datasettet som \emph{ikke}
har en \code{verdi} lik \code{NA} i skåringstabellen (for alle delskalene
som variabelen inngår i).

Funksjonen gir feilmelding dersom noen av verdiene i \code{d} ikke er i
samsvar med skåringstabellen eller dersom skåringstabellen er ugyldig.
}
\examples{
library(tibble)

d_eks = tribble(
  ~pas_id, ~var_a, ~var_b, ~dato,
  146, 1, 2, "2020-05-15",
  211, NA, 3, "2020-08-17"
)

skaaringstabell_eks = tribble(
  ~delskala, ~variabel, ~verdi, ~koeffisient,
  "total", "var_a", 1, 5,
  "total", "var_a", 2, 10,
  "total", "var_a", NA, 7.5,
  "total", "var_b", 1, 2,
  "total", "var_b", 2, 3,
  "total", "var_b", 3, 7,
  "psykisk", "var_a", 1, 0,
  "psykisk", "var_a", 2, 100,
  "psykisk", "var_a", NA, 50
)

skaar_datasett(d_eks, skaaringstabell = skaaringstabell_eks)
}
