% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regn-sumskaar-generelt.R
\name{skaar_datasett_uten_validering}
\alias{skaar_datasett_uten_validering}
\title{Skår garantert gyldig spørreskjema ved hjelp av verditabell}
\usage{
skaar_datasett_uten_validering(d, skaaringstabell)
}
\arguments{
\item{d}{Dataramme/tibble som inneholder kun kolonner med identiske
navn som i \code{skaaringstabell$variabel}. Alle kolonnene må
inneholde numeriske verdier.}

\item{skaaringstabell}{Dataramme/tibble som sier hvordan
\code{d} skal skåres. Må ha fire kolonner,
\code{delskala} (tekst), \code{variabel} (tekst), \code{verdi} (numerisk) og
\code{koeffisient} (numerisk), og det kan bare finnes én rad
per kombinasjon av \code{delskala}, \code{variabel} og \code{verdi}.
Se detaljer nedenfor.}
}
\value{
Tibble som inneholder én eller flere kolonner med
sumskårer. Rekkefølgen på sumskår-kolonnene bestemmes av
rekkefølgen i \code{skaaringstabell$delskala}, og rekkefølgen på radene
er lik som i \code{d}. Hvis en variabel i \code{d} har \code{NA-verdier} uten
tilhørende koeffisient i skåringstabellen, blir sumskåren(e) for
tilhørende delskala(er) satt lik \code{NA}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Skårer et datasett som må være i samsvar med oppgitt verditabell og
gir ut en dataramme/tibble som bare inneholder sumskårene.
}
\details{
Se \code{\link[=skaar_datasett]{skaar_datasett()}} for informasjon om hvordan \code{d} blir skåret
og om hvordan \code{skaaringstabell} skal defineres.
Forskjellen fra den funksjonen er at:
\itemize{
\item \code{d} skal \emph{bare} inneholde variablene nevnt i \code{skaaringstabell$variabel}
(ikke ekstravariabler)
\item \code{d} \emph{må} være i samsvar med \code{skaaringstabell}
\item \code{skaaringstabell} \emph{må} være gyldig
\item utdata fra funksjonen inneholder bare sumskårene, ikke de opphavlige
variablene
}

Ved bruk av den overordnede funksjonen
\code{\link[=skaar_datasett]{skaar_datasett()}} kalles denne funksjonen på etter at
variabler har fått nye namn basert på \code{variabelnavn}-argumentet
i den funksjonen,
skåringstabellen, variabelnavnene og variabelverdiene er validert
(ved hjelp av \code{\link[=sjekk_skaaringstabell]{sjekk_skaaringstabell()}}, \code{\link[=sjekk_variabelnavn]{sjekk_variabelnavn()}} og
\code{\link[=sjekk_variabelverdier]{sjekk_variabelverdier()}}) og alle variabler som ikke
finnes i \code{skaaringstabell$variabel} er filtrert vekk.

Dette er altså funksjonen som gjør \emph{hovedarbeidet} med skåring av
et datasett, etter at alt «bokholderiet» er unnagjort.
}
\keyword{internal}
