% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regn-sumskaar-rand12.R
\name{skaar_rand12}
\alias{skaar_rand12}
\title{Skår RAND-12-/SF-12-spørreskjema}
\usage{
skaar_rand12(d, algoritme, variabelnavn = NULL, godta_manglende = TRUE)
}
\arguments{
\item{d}{Dataramme/tibble som inneholder
spørreskjema-variabler + eventuelt andre variabler.
Variabelnavnene for de tolv spørsmålene skal i utgangspunktet være
\code{rand_1}, \code{rand_2a}, \code{rand_2b}, \code{rand_3a}, \code{rand_3b}, \code{rand_4a},
\code{rand_4b}, \code{rand_5}, \code{rand_6a}, \code{rand_6b}, \code{rand_6c} og \code{rand_7}
(se \code{variabelnavn}).
Se detaljer nedenfor for informasjon om koding.}

\item{algoritme}{Algoritme for skåring av spørreskjemaet.
Foreløpig er det bare mulig å velge \code{"farivar_2007_oblique"},
som gir sumskårer basert på en korrelert (skrå/«oblique»)
faktormodell som definert i
\insertCite{Farivar2007;textual}{rapwhale}.}

\item{variabelnavn}{Hvis \code{d} ikke bruker standard variabelnavn,
kan du her oppgi kobling mellom standard variabelnavn (se ovenfor)
og de brukte variabelnavnene. Bruk syntaksen
\code{c(std_navn_1 = "brukt_navn_1", std_navn_2 = "brukt_navn_2")}.
Navnekobling trenger kun oppgis for de variablene som ikke har
standardnavn. Sett til \code{NULL} hvis alle har standardnavn.}

\item{godta_manglende}{Skal manglende verdier (\code{NA}-verdier) i
spørreskjema-variablene i \code{d} godtas (som standard nei)? Hvis
ikke, blir det gitt ut en feilmelding om det finnes manglende
verdier.}
}
\value{
Datasett likt \code{d}, men med sumskårene \code{rand12_pcs} og
\code{rand12_mcs} lagt til, eventuelt erstattet. Sumskår-kolonnene blir
i utgangspunktet lagt til på slutten av \code{d}.
Hvis \code{d} imidlertid alt innholder en variabel med
navnet \code{rand12_pcs} eller \code{rand12_mcs}, blir denne denne stående
der den er, men overskrevet med nyutregnet sumskår. Det blir i så
fall gitt ut en advarsel.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Regner ut fysisk og psykisk sumskår (PCS og MCS)
for RAND-12-/SF-12-spørreskjema basert valgfri skåringsalgoritme.
Sjekker også at alle variabelnavnene og
verdiene/besvarelsene er gyldige.
}
\details{
Funksjonen regner ut både fysisk og psykisk sumskår (PCS og MCS)
for RAND-12- eller SF-12-spørreskjema og legger disse til datasettet.
Den sjekker også at alle variabelnavnene og verdiene/besvarelsene
er kodet riktig.

Skjemavariablene skal kodes slik at første svaralternativ for hvert
spørsmål har tallkoden 1, andre svaralternativ har tallkoden 2 osv.
For spørsmålene \code{rand_1}, \code{rand_5}, \code{rand_6a} og \code{rand_6b} tilsvarer
altså lav verdi god helse, mens for resten av spørsmålene tilsvarer lav
verdi dårlig helse.

Skåringsalgoritmen i \insertCite{Farivar2007;textual}{rapwhale} er basert på
engelsk versjon av SF-12 versjon 1 (med amerikansk referansepopulasjon),
men den kan også brukes for (og er foreløpig det beste alternativet for)
RAND-12 norsk versjon 1 og SF-12 norsk versjon 1.1 og 1.2.
}
\note{
Funksjonen kan \emph{ikke} brukes for skåring av versjon 2 av SF-12
(dette skjemaet har for noen spørsmål flere mulige svaralternativer).
}
\examples{
library(tibble)

d_eks = tribble(
  ~rand_1, ~rand_2a, ~rand_2b, ~rand_3a, ~rand_3b, ~rand_4a,
  ~rand_4b, ~rand_5, ~rand_6a, ~rand_6b, ~rand_6c, ~rand_7,
  5, 1, 1, 1, 1, 1, 1, 5, 6, 6, 1, 2, # Min PCS12
  1, 3, 3, 2, 2, 2, 2, 1, 1, 1, 6, 5, # Maks PCS12
  5, 3, 3, 1, 1, 1, 1, 5, 6, 6, 1, 2, # Min MCS12
  1, 1, 1, 2, 2, 2, 2, 1, 1, 1, 6, 5, # Maks MCS12
  5, 1, 1, 1, 1, 1, 1, 5, 6, 6, 1, 1
) # Mest negativt svar på alle enkeltspørsmål
skaar_rand12(d_eks, algoritme = "farivar_2007_oblique")
}
\references{
\insertAllCited{}
}
\seealso{
Se \code{\link[=skaar_datasett]{skaar_datasett()}} og underfunksjoner for skåring
av generelle spørreskjemaer.
}
