% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_dd_oqr_csv.R
\name{std_koltype_til_readr_koltype}
\alias{std_koltype_til_readr_koltype}
\title{Konverter variabeltype fra standard format til readrformat}
\usage{
std_koltype_til_readr_koltype(vartype)
}
\arguments{
\item{vartype}{Vektor med variabeltyper oppgitt p책 standardformat.
Mulige typer er: tekst, desimalltall, heltall, boolsk, dato, dato_kl og kl.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{std_koltype_til_readr_koltype()} tar inn en vektor med variabeltyper oppgitt p책 v책rt standardformat.
Returnerer en sammensl책tt tekststreng med gyldige coltypes for bruk i \code{read_*()}-funksjoner fra readr-pakken.
Se \code{read_delim()}
}
\keyword{internal}
