% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymse-funksjoner.R
\name{tab}
\alias{tab}
\title{Tabell inkludert NA-verdier}
\usage{
tab(...)
}
\arguments{
\item{...}{Tilsvarende argumenter som brukes i \code{\link[base:table]{base::table()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Variant av \code{\link[base:table]{base::table()}} som tar med NA-verdiar om dei finst.
}
\examples{
tab(c(rpois(95, 5), rep(NA, 5)))
}
