% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema-kvalreg.R
\name{tema_kvalreg}
\alias{tema_kvalreg}
\title{Tema for bruk i figurer}
\usage{
tema_kvalreg()
}
\value{
Objektet \code{tema} som inneholder temainnstillinger.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Inneholder temainnstillinger som skal brukes i figurer
i årsrapporter.
}
\examples{
# Pakke for å laga figurar
library(ggplot2)

# Sølediagram utan tema_kvalreg()
p = ggplot(mpg, aes(class)) +
  geom_bar()
p

# Søylediagram med tema_kvalreg()
p + tema_kvalreg()
}
