% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_kb_v2.R
\name{tildel_unike_skjemanavn_fra_skjema_id}
\alias{tildel_unike_skjemanavn_fra_skjema_id}
\title{Tildel unike skjemanavn fra tabellnavn}
\usage{
tildel_unike_skjemanavn_fra_skjema_id(kb_std)
}
\arguments{
\item{kb_std}{Kodebok på standardformat}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Gir nye skjemanavn som samsvarer med tabellen variablene ligger i.
Velger første \emph{ledige} navn blant tilgjengelige skjema-id'er.
}
\keyword{internal}
