% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_kb_v2.R
\name{utvid_statusvariabel}
\alias{utvid_statusvariabel}
\title{Utvid statusvariabler til kategorisk}
\usage{
utvid_statusvariabel(kb_mellom)
}
\arguments{
\item{kb_mellom}{Kodebok på mellomformat}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Funksjonen tar statusvariabel fra OQR-register og konverterer den til en
kategorisk variabel. I OQR-struktur har statusvariabel kun et nivå, men
vi vil ha et nivå for hver listeverdi.
Tar inn kodebok og returnerer kodebok med statusvariabler utvidet
}
\keyword{internal}
