% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_kb_v2.R
\name{valider_kb_skjema}
\alias{valider_kb_skjema}
\title{Valider skjema}
\usage{
valider_kb_skjema(kodebok)
}
\arguments{
\item{kodebok}{kodebok p책 standard format.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Et sett med tester for 책 validere kodebok p책 skjemaniv책.
}
\keyword{internal}
