% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_kb_v2.R
\name{valider_kb_struktur}
\alias{valider_kb_struktur}
\title{Valider struktur av kodebok}
\usage{
valider_kb_struktur(kodebok)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Funksjon for å validere struktur av kodebok.
Fullstendig mangelfull per nå, vil lages senere.
}
\keyword{internal}
