% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_kb_v2.R
\name{valider_kb_variabler}
\alias{valider_kb_variabler}
\title{Valider variabler}
\usage{
valider_kb_variabler(kodebok)
}
\arguments{
\item{kodebok}{kodebok p책 standardformat}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Et sett med tester for 책 valider kodebok p책 variabelniv책.
}
\keyword{internal}
