% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_kb_v2.R
\name{valider_kodebok}
\alias{valider_kodebok}
\title{Valider kodebok}
\usage{
valider_kodebok(kodebok)
}
\arguments{
\item{kodebok}{Kodebok p책 standardformat}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Funksjon for 책 validere kodebok for 책 sikre at den er i henhold til v책rt standardformat.
Tar inn kodebok og returnerer ingenting gitt at kodeboken oppfyller kravene til standardformat.
}
\keyword{internal}
