% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/les_kb_v2.R
\name{velg_standardkolonner}
\alias{velg_standardkolonner}
\title{Velg standardkolonner}
\usage{
velg_standardkolonner(kb_mellom)
}
\arguments{
\item{kb_mellom}{Kodebok på mellomformat}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Velger ut standardkolonner tilsvarende vårt standardformat. Kolonnene
hentes ut i standard rekkefølge.
}
\keyword{internal}
