% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapwhale-styler-funksjonar.R
\name{wrap_curly_rapwhale}
\alias{wrap_curly_rapwhale}
\title{Wrap a statement in curly braces}
\usage{
wrap_curly_rapwhale(pd, indent_by, key_token, space_after = 1)
}
\arguments{
\item{pd}{A parse table.}

\item{indent_by}{The amount of spaces used to indent an expression in curly braces.
Used for unindention.}

\item{key_token}{The token that comes right before the token that contains the expression
to be wrapped (ignoring comments).
For if and while loops, this is the closing "')'",
for a for-loop it's "forcond".}

\item{space_after}{How many spaces should be inserted after the closing brace.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Wrap a statement in curly braces.
}
\details{
Lik \link[styler:wrap_multiline_curly]{styler::wrap_multiline_curly},
men der alle uttrykk får krøllparentesar (ikkje berre multi-line).
}
\keyword{internal}
