% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapwhale-styler-funksjonar.R
\name{wrap_if_else_while_for_fun_in_curly_rapwhale}
\alias{wrap_if_else_while_for_fun_in_curly_rapwhale}
\title{Wrap if-else, while and for statements in curly braces}
\usage{
wrap_if_else_while_for_fun_in_curly_rapwhale(pd, indent_by = 2)
}
\arguments{
\item{pd}{A parse table.}

\item{indent_by}{The amount of spaces used to indent an expression in curly braces.
Used for unindention.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Wrap statements in curly braces if it is not already wrapped in a such.
}
\details{
Lik \link[styler:wrap_if_else_while_for_function_multi_line_in_curly]{styler::wrap_if_else_while_for_function_multi_line_in_curly},
men der alle uttrykk får krøllparentesar (ikkje berre multi-line).
}
\keyword{internal}
