% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figur-funksjoner.R
\name{lag_fig_soyle}
\alias{lag_fig_soyle}
\title{Lag søylediagram}
\usage{
lag_fig_soyle(d, x, y, flip = TRUE, ...)
}
\arguments{
\item{d}{Datasett som inkluderer dei variablane som skal brukast søylediagrammet.}

\item{x}{Variabel for x-aksen - Vanlegvis er dette ein kategorisk variabel.}

\item{y}{Variabel for y-aksen - Ein kontinuerleg variabel.
For prosent, bruk \code{\link[=lag_fig_soyle_prosent]{lag_fig_soyle_prosent()}}.}

\item{flip}{Logisk variabel som seier om plottet skal flippast. Standard verdi = \code{TRUE}.}

\item{...}{Ekstra argument som vert gjeve vidare til \code{\link[=lag_fig_soyle_grunnplott]{lag_fig_soyle_grunnplott()}}
(og så vidare igjen til \code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}}).}
}
\value{
Eit søylediagram som ggplot-objekt.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Funksjon for å laga søylediagram
}
\examples{
library(tibble)

d = tibble(gruppe = c("a", "b", "c"), verdi = c(2.6, 2.1, 3.2))
lag_fig_soyle(d, gruppe, verdi)
lag_fig_soyle(d, gruppe, verdi, flip = FALSE)
}
